Bayesian Sampling of Mixture Autoregressive Models

Samples parameters of a mixture autoregressive model from respective posterior distributions.

bayes_mixAR(y, model, fix_shift = FALSE, a = .2, c = 2, tau, nsim, burnin)

Arguments

y
a time series (currently a numeric vector).

model
an object of class mixAR. Currently only handles mixARGaussian objects.

fix_shift
Should shift be kept as fixed? If FALSE (default), shift is sampled.

a, c
numeric hyperparameters, default values are taken from Richardson and Green (1997).

tau
numeric vector of length g, the number of components in the mixture. Tuning parameter for M-H move in updating AR parameters. If length(tau) is 1, same tuning parameter is taken for all components.

nsim
numeric, the number of iterations.

burnin
numeric, the number of iterations taken as burn-in period.

Details

For details see Ravagli and Boshnakov (2019).

Value

a list with following elements:

mix_weights

a g columns matrix with samples from the posterior distributions of the mixing weights.

scale

a g columns matrix with samples from posterior distributions of scale parameters.

precision

a g columns matrix with samples from posterior distributions of precision parameters, defined as 1 / (scale ^ 2).

shift

a g columns matrix with samples from posterior distributions of shift parameters, namely phi_k0.

mu

a g columns matrix with samples from posterior distributions of component means, calculated as phi_k0 / (1 - phi_k1 - phi_k2 - ...).

ARcoeff

a list which elements are matrices, one for each AR component in the mixture.

acc_rate

numeric vector, the acceptance rate for M-H moves.

n_samp

the sample size, calculated as nsim - burnin.

LatentZ

the latest Z variables drawn (for utility only).

n_comp

the number of components in the mixture.

fix_shift

same as input, whether the shift parameter was kept fixed or not.

References

Richardson S, Green PJ (1997). “On Bayesian Analysis of Mixtures with an Unknown Number of Components.” J. R. Stat. Soc., Ser. B, Stat. Methodol. , 59(4), 731-792.

Ravagli D, Boshnakov GN (2019). “Bayesian analysis of mixture autoregressive models covering the complete parameter space.” (in preparation).

Examples

prob <- c(0.5, 0.5)
sigma <- c(1, 2)
ar <- list(-0.5, 1)

model <- new("mixARGaussian", prob = prob, scale = sigma, arcoef = ar)

## MAR(1,1) model
y <- mixAR_sim(model, 300, rep(0, max(model@order)))

bayes_mixAR(y, model, fix_shift = FALSE, tau = c(.15,.25), nsim = 20, burnin = 10)

#> $mix_weights
#>            [,1]      [,2]
#>  [1,] 0.5944429 0.4055571
#>  [2,] 0.6630588 0.3369412
#>  [3,] 0.6865778 0.3134222
#>  [4,] 0.5785291 0.4214709
#>  [5,] 0.5991506 0.4008494
#>  [6,] 0.5825313 0.4174687
#>  [7,] 0.5437193 0.4562807
#>  [8,] 0.6336777 0.3663223
#>  [9,] 0.6575754 0.3424246
#> [10,] 0.6130074 0.3869926
#> 
#> $scale
#>           [,1]     [,2]
#>  [1,] 1.062397 1.704061
#>  [2,] 1.088883 1.538802
#>  [3,] 1.064843 1.525148
#>  [4,] 1.070298 1.704299
#>  [5,] 0.969904 1.584791
#>  [6,] 1.031180 1.564473
#>  [7,] 1.112665 1.680976
#>  [8,] 1.236402 1.728616
#>  [9,] 1.166900 1.702312
#> [10,] 1.218137 1.513038
#> 
#> $precision
#>            [,1]      [,2]
#>  [1,] 0.8859852 0.3443735
#>  [2,] 0.8434069 0.4223129
#>  [3,] 0.8819196 0.4299083
#>  [4,] 0.8729517 0.3442772
#>  [5,] 1.0630225 0.3981587
#>  [6,] 0.9404400 0.4085677
#>  [7,] 0.8077390 0.3538971
#>  [8,] 0.6541549 0.3346592
#>  [9,] 0.7344001 0.3450815
#> [10,] 0.6739192 0.4368176
#> 
#> $shift
#>              [,1]        [,2]
#>  [1,]  0.02525309  0.15356254
#>  [2,]  0.07580856 -0.28872949
#>  [3,]  0.06113259 -0.73642494
#>  [4,]  0.13847197 -0.33180628
#>  [5,]  0.17312270 -0.25539391
#>  [6,]  0.06949121 -0.28572613
#>  [7,]  0.05998298 -0.21099859
#>  [8,] -0.03203843 -0.26005817
#>  [9,]  0.12317250 -0.18663956
#> [10,] -0.17665500 -0.09728805
#> 
#> $mu
#>              [,1]      [,2]
#>  [1,]  0.01731888 1.4793558
#>  [2,]  0.05199043 2.5282453
#>  [3,]  0.04192548 6.4484681
#>  [4,]  0.09407911 2.9054451
#>  [5,]  0.11762113 2.2363440
#>  [6,]  0.04721296 2.5019465
#>  [7,]  0.04075298 1.8475986
#>  [8,] -0.02237026 2.2771863
#>  [9,]  0.08167014 1.6343000
#> [10,] -0.11713197 0.8518979
#> 
#> $ARcoeff
#> $ARcoeff$Component_1
#>             [,1]
#>  [1,] -0.4581252
#>  [2,] -0.4581252
#>  [3,] -0.4718674
#>  [4,] -0.4718674
#>  [5,] -0.4718674
#>  [6,] -0.4718674
#>  [7,] -0.4321888
#>  [8,] -0.5081706
#>  [9,] -0.5081706
#> [10,] -0.5081706
#> 
#> $ARcoeff$Component_2
#>           [,1]
#>  [1,] 1.114202
#>  [2,] 1.114202
#>  [3,] 1.114202
#>  [4,] 1.114202
#>  [5,] 1.114202
#>  [6,] 1.114202
#>  [7,] 1.114202
#>  [8,] 1.114202
#>  [9,] 1.114202
#> [10,] 1.114202
#> 
#> 
#> $acc_rate
#> [1] 0.3 0.1
#> 
#> $n_samp
#> [1] 10
#> 
#> $LatentZ
#>        latent Z  1 latent Z  2
#>   [1,]           1           0
#>   [2,]           0           1
#>   [3,]           0           1
#>   [4,]           1           0
#>   [5,]           0           1
#>   [6,]           1           0
#>   [7,]           1           0
#>   [8,]           1           0
#>   [9,]           1           0
#>  [10,]           0           1
#>  [11,]           0           1
#>  [12,]           1           0
#>  [13,]           1           0
#>  [14,]           0           1
#>  [15,]           1           0
#>  [16,]           1           0
#>  [17,]           1           0
#>  [18,]           1           0
#>  [19,]           1           0
#>  [20,]           1           0
#>  [21,]           1           0
#>  [22,]           1           0
#>  [23,]           1           0
#>  [24,]           1           0
#>  [25,]           1           0
#>  [26,]           1           0
#>  [27,]           1           0
#>  [28,]           0           1
#>  [29,]           0           1
#>  [30,]           1           0
#>  [31,]           1           0
#>  [32,]           1           0
#>  [33,]           0           1
#>  [34,]           1           0
#>  [35,]           0           1
#>  [36,]           1           0
#>  [37,]           0           1
#>  [38,]           0           1
#>  [39,]           1           0
#>  [40,]           1           0
#>  [41,]           1           0
#>  [42,]           0           1
#>  [43,]           1           0
#>  [44,]           0           1
#>  [45,]           1           0
#>  [46,]           0           1
#>  [47,]           1           0
#>  [48,]           1           0
#>  [49,]           1           0
#>  [50,]           1           0
#>  [51,]           1           0
#>  [52,]           0           1
#>  [53,]           1           0
#>  [54,]           1           0
#>  [55,]           1           0
#>  [56,]           0           1
#>  [57,]           0           1
#>  [58,]           0           1
#>  [59,]           0           1
#>  [60,]           1           0
#>  [61,]           1           0
#>  [62,]           1           0
#>  [63,]           0           1
#>  [64,]           1           0
#>  [65,]           0           1
#>  [66,]           0           1
#>  [67,]           0           1
#>  [68,]           1           0
#>  [69,]           1           0
#>  [70,]           1           0
#>  [71,]           1           0
#>  [72,]           0           1
#>  [73,]           1           0
#>  [74,]           0           1
#>  [75,]           0           1
#>  [76,]           0           1
#>  [77,]           1           0
#>  [78,]           1           0
#>  [79,]           1           0
#>  [80,]           1           0
#>  [81,]           0           1
#>  [82,]           1           0
#>  [83,]           1           0
#>  [84,]           1           0
#>  [85,]           0           1
#>  [86,]           1           0
#>  [87,]           0           1
#>  [88,]           0           1
#>  [89,]           0           1
#>  [90,]           1           0
#>  [91,]           1           0
#>  [92,]           1           0
#>  [93,]           1           0
#>  [94,]           0           1
#>  [95,]           1           0
#>  [96,]           1           0
#>  [97,]           1           0
#>  [98,]           0           1
#>  [99,]           1           0
#> [100,]           0           1
#> [101,]           0           1
#> [102,]           0           1
#> [103,]           1           0
#> [104,]           0           1
#> [105,]           0           1
#> [106,]           0           1
#> [107,]           0           1
#> [108,]           0           1
#> [109,]           0           1
#> [110,]           1           0
#> [111,]           0           1
#> [112,]           0           1
#> [113,]           1           0
#> [114,]           1           0
#> [115,]           1           0
#> [116,]           1           0
#> [117,]           1           0
#> [118,]           1           0
#> [119,]           0           1
#> [120,]           1           0
#> [121,]           1           0
#> [122,]           1           0
#> [123,]           0           1
#> [124,]           0           1
#> [125,]           1           0
#> [126,]           1           0
#> [127,]           1           0
#> [128,]           0           1
#> [129,]           1           0
#> [130,]           0           1
#> [131,]           1           0
#> [132,]           0           1
#> [133,]           0           1
#> [134,]           1           0
#> [135,]           1           0
#> [136,]           1           0
#> [137,]           1           0
#> [138,]           1           0
#> [139,]           0           1
#> [140,]           1           0
#> [141,]           0           1
#> [142,]           0           1
#> [143,]           0           1
#> [144,]           0           1
#> [145,]           1           0
#> [146,]           0           1
#> [147,]           1           0
#> [148,]           0           1
#> [149,]           1           0
#> [150,]           0           1
#> [151,]           0           1
#> [152,]           0           1
#> [153,]           1           0
#> [154,]           1           0
#> [155,]           1           0
#> [156,]           0           1
#> [157,]           0           1
#> [158,]           1           0
#> [159,]           1           0
#> [160,]           1           0
#> [161,]           1           0
#> [162,]           1           0
#> [163,]           1           0
#> [164,]           0           1
#> [165,]           0           1
#> [166,]           0           1
#> [167,]           0           1
#> [168,]           1           0
#> [169,]           1           0
#> [170,]           1           0
#> [171,]           1           0
#> [172,]           1           0
#> [173,]           0           1
#> [174,]           1           0
#> [175,]           1           0
#> [176,]           1           0
#> [177,]           1           0
#> [178,]           1           0
#> [179,]           1           0
#> [180,]           1           0
#> [181,]           1           0
#> [182,]           0           1
#> [183,]           1           0
#> [184,]           1           0
#> [185,]           1           0
#> [186,]           0           1
#> [187,]           1           0
#> [188,]           1           0
#> [189,]           1           0
#> [190,]           1           0
#> [191,]           1           0
#> [192,]           0           1
#> [193,]           1           0
#> [194,]           0           1
#> [195,]           1           0
#> [196,]           0           1
#> [197,]           1           0
#> [198,]           1           0
#> [199,]           1           0
#> [200,]           1           0
#> [201,]           0           1
#> [202,]           0           1
#> [203,]           0           1
#> [204,]           0           1
#> [205,]           1           0
#> [206,]           0           1
#> [207,]           1           0
#> [208,]           1           0
#> [209,]           0           1
#> [210,]           1           0
#> [211,]           1           0
#> [212,]           0           1
#> [213,]           1           0
#> [214,]           1           0
#> [215,]           0           1
#> [216,]           0           1
#> [217,]           0           1
#> [218,]           1           0
#> [219,]           0           1
#> [220,]           1           0
#> [221,]           1           0
#> [222,]           1           0
#> [223,]           1           0
#> [224,]           1           0
#> [225,]           1           0
#> [226,]           1           0
#> [227,]           1           0
#> [228,]           0           1
#> [229,]           0           1
#> [230,]           0           1
#> [231,]           1           0
#> [232,]           1           0
#> [233,]           1           0
#> [234,]           0           1
#> [235,]           0           1
#> [236,]           0           1
#> [237,]           1           0
#> [238,]           1           0
#> [239,]           0           1
#> [240,]           0           1
#> [241,]           1           0
#> [242,]           1           0
#> [243,]           1           0
#> [244,]           1           0
#> [245,]           1           0
#> [246,]           0           1
#> [247,]           1           0
#> [248,]           0           1
#> [249,]           0           1
#> [250,]           1           0
#> [251,]           1           0
#> [252,]           0           1
#> [253,]           1           0
#> [254,]           1           0
#> [255,]           1           0
#> [256,]           1           0
#> [257,]           1           0
#> [258,]           0           1
#> [259,]           1           0
#> [260,]           1           0
#> [261,]           1           0
#> [262,]           1           0
#> [263,]           1           0
#> [264,]           1           0
#> [265,]           1           0
#> [266,]           1           0
#> [267,]           0           1
#> [268,]           1           0
#> [269,]           0           1
#> [270,]           1           0
#> [271,]           1           0
#> [272,]           0           1
#> [273,]           1           0
#> [274,]           0           1
#> [275,]           1           0
#> [276,]           1           0
#> [277,]           1           0
#> [278,]           1           0
#> [279,]           1           0
#> [280,]           1           0
#> [281,]           1           0
#> [282,]           0           1
#> [283,]           1           0
#> [284,]           0           1
#> [285,]           1           0
#> [286,]           1           0
#> [287,]           1           0
#> [288,]           1           0
#> [289,]           1           0
#> [290,]           0           1
#> [291,]           1           0
#> [292,]           1           0
#> [293,]           0           1
#> [294,]           1           0
#> [295,]           0           1
#> [296,]           1           0
#> [297,]           1           0
#> [298,]           1           0
#> [299,]           1           0
#> 
#> $ncomp
#> [1] 2
#> 
#> $fix_shift
#> [1] FALSE
#>