Functions for the standard normal distribution

The noise distributions are specified by a list of functions for the density, quantiles, etc. This object demonstrates this for the standard normal distribution.

Usage

dist_norm

Format

This is a list of functions or names of functions for calculations related to the standard normal distribution. Currently it has elements with the following names: "pdf", "cdf", "rand", "logpdf", "Fscore", "xFscore", "Parscore", "get_param", "set_param", "any_param", "show" .

Details

dist_norm may be used to specify the noise distribution for MixAR models. It can be used as a template if other distributions are needed, see also fdist_stdnorm.

See also

Examples

dist_norm
#> $pdf
#> [1] "dnorm"
#> 
#> $cdf
#> [1] "pnorm"
#> 
#> $rand
#> [1] "rnorm"
#> 
#> $logpdf
#> function (x) 
#> log(dnorm(x))
#> <bytecode: 0x56104ad93e18>
#> <environment: 0x56104d50f2b8>
#> 
#> $Fscore
#> function (x) 
#> -x
#> <bytecode: 0x56104ad946a0>
#> <environment: 0x56104d50f2b8>
#> 
#> $xFscore
#> function (x) 
#> -x^2
#> <bytecode: 0x56104ad94cc0>
#> <environment: 0x56104d50f2b8>
#> 
#> $Parscore
#> function () 
#> 0
#> <bytecode: 0x56104ad8f590>
#> <environment: 0x56104d50f2b8>
#> 
#> $get_param
#> function () 
#> numeric()
#> <bytecode: 0x56104ad8fa98>
#> <environment: 0x56104d50f2b8>
#> 
#> $set_param
#> function () 
#> stop("Standard normal has no parameters to set.")
#> <bytecode: 0x56104ad90048>
#> <environment: 0x56104d50f2b8>
#> 
#> $any_param
#> function () 
#> param_flag
#> <bytecode: 0x56104ad905c0>
#> <environment: 0x56104d50f2b8>
#> 
#> $show
#> function () 
#> "Standard normal distribution"
#> <bytecode: 0x56104ad90d68>
#> <environment: 0x56104d50f2b8>
#> 
dist_norm$pdf
#> [1] "dnorm"
dist_norm$cdf
#> [1] "pnorm"