A posteriori relabelling of a Markov chain

Takes the output from a MCMC simulation of parameters of a mixture, and detects whether labels switch has occured while sampling, using the method by Celeux (2000).

label_switch(x, m)

Arguments

x
output from an MCMC sampling of a mixture. A matrix, each column corresponds to one component of the mixture.

m
the number of observations in the sample that will be used to initialise the algorithm. m~100 is recommended.

Details

Function can be directly executed when x is one of mix_weights, scale, precision, shift or mu from bayes_mixAR output. ARcoeff cannot be input as it is, but element from the list may be used.

Value

A list of 2:

x

The input matrix, with adjusted labels

true_perm

The "true" permutation at each iteration.

References

Celeux G (2000). Bayesian Inference of Mixture: The Label Switching Problem.. Payne R., Green P. (eds) COMPSTAT. Physica, Heidelberg.

Note

There is no absolute choice on what x should be to obtain the "true" permutation at any given point. User is subject to make the most suitable choice, given output of their MCMC.

See also

Examples

model <- new("MixARGaussian",
             prob   = exampleModels$WL_At@prob,      # c(0.5, 0.5)
             scale  = exampleModels$WL_At@scale,     # c(1, 2)        
             arcoef = exampleModels$WL_At@arcoef@a ) # list(-0.5, 1.1)

y <- mixAR_sim(model, n = 300, init = rep(0, which.max(model@order)))

## just examples, use larger numbers in practice
nsim <- 30   # 200
burnin <- 10  # 100
x <- bayes_mixAR(y, model, fix_shift = FALSE, tau = c(.15, .25),
              nsim = nsim, burnin = burnin)

label_switch(x$mix_weights, m = 5)

#> $x
#>            [,1]      [,2]
#>  [1,] 0.3840002 0.6159998
#>  [2,] 0.3878427 0.6121573
#>  [3,] 0.4526472 0.5473528
#>  [4,] 0.4023388 0.5976612
#>  [5,] 0.4515958 0.5484042
#>  [6,] 0.4095133 0.5904867
#>  [7,] 0.4955273 0.5044727
#>  [8,] 0.4929660 0.5070340
#>  [9,] 0.4693955 0.5306045
#> [10,] 0.4734032 0.5265968
#> [11,] 0.3712098 0.6287902
#> [12,] 0.3656195 0.6343805
#> [13,] 0.3789003 0.6210997
#> [14,] 0.4461685 0.5538315
#> [15,] 0.4409884 0.5590116
#> [16,] 0.4647830 0.5352170
#> [17,] 0.4763755 0.5236245
#> [18,] 0.4526394 0.5473606
#> [19,] 0.4372181 0.5627819
#> [20,] 0.4665436 0.5334564
#> 
#> $true_perm
#>       [,1] [,2]
#>  [1,]    1    2
#>  [2,]    1    2
#>  [3,]    1    2
#>  [4,]    1    2
#>  [5,]    1    2
#>  [6,]    1    2
#>  [7,]    1    2
#>  [8,]    1    2
#>  [9,]    1    2
#> [10,]    1    2
#> [11,]    1    2
#> [12,]    1    2
#> [13,]    1    2
#> [14,]    1    2
#> [15,]    1    2
#>