Simulation experiments with MixAR models

Perform simulation experiments with MixAR models.

mixARExperiment(model, imodel = NULL, simargs = NULL, estargs = NULL, fix, ...)

Arguments

model
the underlying model, an object inheriting from "MixAR", see 'Details'.

imodel
initial model, an object inheriting from "MixAR", see 'Details'.

simargs
additional arguments for the simulation function, a list, see 'Details'.

estargs
additional arguments for the estimation function, a list, see 'Details'.

fix
which arguments to keep fixed during estimation; currently if fix = "shift" the shift parameters are kept fixed, otherwise all parameters are estimated.

...
additional arguments to pass on to simuExperiment.

Details

mixARExperiment is a wrapper for simuExperiment.

The simulation function is mixARsim and the estimation function is fit_mixAR. The simulation function gets model as first argument. The first argument of the estimation function is imodel, if supplied, or model otherwise. Some appropriate defaults for other arguments are also supplied. Arguments simargs and estargs are needed only to override defaults.

If fix = "shift", the shifts components are not estimated and are kept fixed. This may be useful when fitting the model to differenced series, for example.

mixARExperiment sets keep and summary_fun to some defaults (currently mixAR:::.fsumA for the latter).

Value

A list with one or more elements, depending on the arguments ... passed on to simuExperiment.

Summary

a summary of the experiment, by default sample means and standard deviations of the estimates. This is a list of MixAR models.

Raw

A list of the estimated models.

Note

This is an initial version of this function, some of its arguments may change. Additional elements may be made available in the returned value, so refer to those by name, not with a numeric index.

See also

Examples

exampleModels$WL_II

#> (To see the internal structure of the object, use function 'str'.)
#> 
#> An object of class "MixARGaussian"
#> Number of components: 3 
#>        prob shift scale order ar_1 ar_2
#> Comp_1 0.4    5     1     2    0.9 -0.6
#> Comp_2 0.3    0     1     2   -0.7  0.0
#> Comp_3 0.3   -5     5     2    0.0  0.8
#> 
#> Distributions of the error components:
#> 	standard Gaussian
#> 

## N of course is not realistic here but unswitch() sometimes drops all
## as unsuccessful (see fsummary()) and error pops up. Need to guard
## against this.
set.seed(1234)
mixARExperiment(exampleModels$WL_II, N = 10)

#> sa$init is:  0 0 
#> Initial vallogf:  -555.8538 
#> Final niter:  11 	vallogf:  -551.8867 
#> Initial vallogf:  -525.9453 
#> Final niter:  11 	vallogf:  -518.5515 
#> Initial vallogf:  -540.2367 
#> Final niter:  11 	vallogf:  -533.9115 
#> Initial vallogf:  -561.8359 
#> Final niter:  11 	vallogf:  -551.7269 
#> Initial vallogf:  -564.1359 
#> Final niter:  11 	vallogf:  -557.3694 
#> Initial vallogf:  -526.1191 
#> Final niter:  11 	vallogf:  -517.3952 
#> Initial vallogf:  -536.4386 
#> Final niter:  11 	vallogf:  -527.288 
#> Initial vallogf:  -572.0168 
#> Final niter:  11 	vallogf:  -564.0925 
#> Initial vallogf:  -569.1482 
#> Final niter:  11 	vallogf:  -563.2507 
#> Initial vallogf:  -564.3853 
#> Final niter:  11 	vallogf:  -555.5492 
#> [1] 10 21  3 10 53  1
#> [1] 10 21  3 10 53  1
#> [1] 10 21  3 10 53  1
#> [1] 10 21  3 10 53  1
#> [1] 10 21  3 10 53  1
#> [1] 10 21  3 10 53  1
#> [1] 10 21  3 10 53  1
#> [1] 10 21  3 10 53  1
#> [1] 10 21  3 10 53  1
#> [1] 10 21  3 10 53  1

#> $Summary
#> $Summary$mean
#> (To see the internal structure of the object, use function 'str'.)
#> 
#> An object of class "MixARGaussian"
#> Number of components: 3 
#>        prob      shift       scale    order ar_1        ar_2      
#> Comp_1 0.3863061  4.97365098 0.994058   2    0.90196708 -0.5999090
#> Comp_2 0.3067055 -0.02257295 1.021391   2   -0.71204858  0.0077286
#> Comp_3 0.3069883 -5.13523149 4.567927   2    0.01004161  0.7613427
#> 
#> Distributions of the error components:
#> 	standard Gaussian
#> 
#> 
#> $Summary$sd
#> (To see the internal structure of the object, use function 'str'.)
#> 
#> An object of class "MixARGaussian"
#> Number of components: 3 
#>        prob       shift     scale      order ar_1       ar_2      
#> Comp_1 0.04551702 0.1255297 0.08789779   2   0.02123861 0.01576425
#> Comp_2 0.03780447 0.2024097 0.11423832   2   0.01645179 0.02720160
#> Comp_3 0.03316003 0.5381740 0.42290391   2   0.06428824 0.09026697
#> 
#> Distributions of the error components:
#> 	standard Gaussian
#> 
#> 
#> $Summary$meanUnswitched
#> (To see the internal structure of the object, use function 'str'.)
#> 
#> An object of class "MixARGaussian"
#> Number of components: 3 
#>        prob      shift       scale    order ar_1        ar_2      
#> Comp_1 0.3067055 -0.02257295 1.021391   2   -0.71204858  0.0077286
#> Comp_2 0.3069883 -5.13523149 4.567927   2    0.01004161  0.7613427
#> Comp_3 0.3863061  4.97365098 0.994058   2    0.90196708 -0.5999090
#> 
#> Distributions of the error components:
#> 	standard Gaussian
#> 
#> 
#> $Summary$sdUnswitched
#> (To see the internal structure of the object, use function 'str'.)
#> 
#> An object of class "MixARGaussian"
#> Number of components: 3 
#>        prob       shift     scale      order ar_1       ar_2      
#> Comp_1 0.03780447 0.2024097 0.11423832   2   0.01645179 0.02720160
#> Comp_2 0.03316003 0.5381740 0.42290391   2   0.06428824 0.09026697
#> Comp_3 0.04551702 0.1255297 0.08789779   2   0.02123861 0.01576425
#> 
#> Distributions of the error components:
#> 	standard Gaussian
#> 
#> 
#> 

mixARExperiment(exampleModels$WL_II, N = 10, raw = TRUE)

#> sa$init is:  0 0 
#> Initial vallogf:  -535.4214 
#> Final niter:  11 	vallogf:  -526.6723 
#> Initial vallogf:  -568.5251 
#> Final niter:  11 	vallogf:  -562.0025 
#> Initial vallogf:  -560.8672 
#> Final niter:  11 	vallogf:  -552.1119 
#> Initial vallogf:  -554.365 
#> Final niter:  11 	vallogf:  -549.1815 
#> Initial vallogf:  -566.1517 
#> Final niter:  11 	vallogf:  -558.3798 
#> Initial vallogf:  -548.9191 
#> Final niter:  11 	vallogf:  -543.6229 
#> Initial vallogf:  -535.4779 
#> Final niter:  11 	vallogf:  -530.7428 
#> Initial vallogf:  -550.5629 
#> Final niter:  11 	vallogf:  -547.9877 
#> Initial vallogf:  -573.3305 
#> Final niter:  11 	vallogf:  -566.2851 
#> Initial vallogf:  -558.6766 
#> Final niter:  11 	vallogf:  -544.8742 
#> [1]  4 31  6  1 55  1
#> [1] 10 26  6  0 54  2
#> [1]  8 29  6  0 54  1
#> [1]  8 28  6  0 54  2
#> [1] 12 25  6  0 54  1
#> [1]  6 29  6  0 56  1
#> [1] 10 25  6  0 54  3
#> [1]  3 33  6  0 54  2
#> [1] 12 23  6  0 54  3
#> [1] 16 21  6  0 54  1

#> $Summary
#> $Summary$mean
#> (To see the internal structure of the object, use function 'str'.)
#> 
#> An object of class "MixARGaussian"
#> Number of components: 3 
#>        prob      shift       scale     order ar_1        ar_2        
#> Comp_1 0.3787856  5.03208009 0.9493346   2    0.90744122 -0.607663556
#> Comp_2 0.2913316 -0.06428409 0.9819901   2   -0.69686899 -0.005645078
#> Comp_3 0.3298828 -4.75006119 4.6640449   2   -0.03276926  0.808600588
#> 
#> Distributions of the error components:
#> 	standard Gaussian
#> 
#> 
#> $Summary$sd
#> (To see the internal structure of the object, use function 'str'.)
#> 
#> An object of class "MixARGaussian"
#> Number of components: 3 
#>        prob       shift     scale     order ar_1        ar_2      
#> Comp_1 0.03317187 0.1533820 0.1023591   2   0.006878745 0.01096513
#> Comp_2 0.02219037 0.1417895 0.1053999   2   0.020845090 0.01963965
#> Comp_3 0.03269626 0.8883508 0.5911968   2   0.063097220 0.04599522
#> 
#> Distributions of the error components:
#> 	standard Gaussian
#> 
#> 
#> $Summary$meanUnswitched
#> (To see the internal structure of the object, use function 'str'.)
#> 
#> An object of class "MixARGaussian"
#> Number of components: 3 
#>        prob      shift       scale     order ar_1        ar_2        
#> Comp_1 0.2913316 -0.06428409 0.9819901   2   -0.69686899 -0.005645078
#> Comp_2 0.3298828 -4.75006119 4.6640449   2   -0.03276926  0.808600588
#> Comp_3 0.3787856  5.03208009 0.9493346   2    0.90744122 -0.607663556
#> 
#> Distributions of the error components:
#> 	standard Gaussian
#> 
#> 
#> $Summary$sdUnswitched
#> (To see the internal structure of the object, use function 'str'.)
#> 
#> An object of class "MixARGaussian"
#> Number of components: 3 
#>        prob       shift     scale     order ar_1        ar_2      
#> Comp_1 0.02219037 0.1417895 0.1053999   2   0.020845090 0.01963965
#> Comp_2 0.03269626 0.8883508 0.5911968   2   0.063097220 0.04599522
#> Comp_3 0.03317187 0.1533820 0.1023591   2   0.006878745 0.01096513
#> 
#> Distributions of the error components:
#> 	standard Gaussian
#> 
#> 
#> 
#> $Raw
#> $Raw[[1]]
#> (To see the internal structure of the object, use function 'str'.)
#> 
#> An object of class "MixARGaussian"
#> Number of components: 3 
#>        prob      shift      scale    order ar_1        ar_2        
#> Comp_1 0.4405257  4.8667357 1.089210   2    0.92215036 -0.608349180
#> Comp_2 0.2909894 -0.1761727 1.131249   2   -0.69879166  0.002115683
#> Comp_3 0.2684848 -4.2341127 3.525107   2    0.04938571  0.793800787
#> 
#> Distributions of the error components:
#> 	standard Gaussian
#> 
#> 
#> $Raw[[2]]
#> (To see the internal structure of the object, use function 'str'.)
#> 
#> An object of class "MixARGaussian"
#> Number of components: 3 
#>        prob      shift       scale     order ar_1        ar_2       
#> Comp_1 0.3437826  4.98809328 1.0442099   2    0.90917379 -0.59893019
#> Comp_2 0.3099672  0.01903451 0.9681159   2   -0.72212279  0.01898886
#> Comp_3 0.3462502 -3.44166594 5.1861298   2   -0.06375311  0.75260427
#> 
#> Distributions of the error components:
#> 	standard Gaussian
#> 
#> 
#> $Raw[[3]]
#> (To see the internal structure of the object, use function 'str'.)
#> 
#> An object of class "MixARGaussian"
#> Number of components: 3 
#>        prob      shift       scale     order ar_1         ar_2        
#> Comp_1 0.3374142  4.97523342 0.8370702   2    0.904970380 -0.599461987
#> Comp_2 0.3019611  0.07057867 0.8843041   2   -0.686226693  0.001485138
#> Comp_3 0.3606247 -6.41349346 4.4520265   2    0.003898724  0.838798839
#> 
#> Distributions of the error components:
#> 	standard Gaussian
#> 
#> 
#> $Raw[[4]]
#> (To see the internal structure of the object, use function 'str'.)
#> 
#> An object of class "MixARGaussian"
#> Number of components: 3 
#>        prob      shift        scale     order ar_1        ar_2        
#> Comp_1 0.3440990  4.913594384 0.8981187   2    0.90405711 -0.630729940
#> Comp_2 0.3157215  0.002650065 0.9465215   2   -0.68059182 -0.007590315
#> Comp_3 0.3401795 -5.491522074 4.7300259   2    0.03155174  0.843783743
#> 
#> Distributions of the error components:
#> 	standard Gaussian
#> 
#> 
#> $Raw[[5]]
#> (To see the internal structure of the object, use function 'str'.)
#> 
#> An object of class "MixARGaussian"
#> Number of components: 3 
#>        prob      shift      scale     order ar_1        ar_2       
#> Comp_1 0.3996076  5.1912035 0.8569677   2    0.90513993 -0.60282359
#> Comp_2 0.2518384 -0.1703149 1.0740823   2   -0.71441249 -0.02112711
#> Comp_3 0.3485540 -4.3952546 5.4089345   2    0.01601168  0.83278263
#> 
#> Distributions of the error components:
#> 	standard Gaussian
#> 
#> 
#> $Raw[[6]]
#> (To see the internal structure of the object, use function 'str'.)
#> 
#> An object of class "MixARGaussian"
#> Number of components: 3 
#>        prob      shift       scale     order ar_1        ar_2       
#> Comp_1 0.3984618  5.06468883 1.0944019   2    0.90443546 -0.61002085
#> Comp_2 0.3033428  0.02337516 0.9389744   2   -0.72477285  0.01683848
#> Comp_3 0.2981953 -5.14214304 4.0439948   2   -0.08124958  0.86325465
#> 
#> Distributions of the error components:
#> 	standard Gaussian
#> 
#> 
#> $Raw[[7]]
#> (To see the internal structure of the object, use function 'str'.)
#> 
#> An object of class "MixARGaussian"
#> Number of components: 3 
#>        prob      shift       scale     order ar_1        ar_2        
#> Comp_1 0.4074589  5.11492417 0.9858900   2    0.91459344 -0.590260590
#> Comp_2 0.2932422 -0.02124257 0.8293895   2   -0.70108930 -0.009203199
#> Comp_3 0.2992989 -4.87509117 4.3723318   2   -0.06956142  0.736725169
#> 
#> Distributions of the error components:
#> 	standard Gaussian
#> 
#> 
#> $Raw[[8]]
#> (To see the internal structure of the object, use function 'str'.)
#> 
#> An object of class "MixARGaussian"
#> Number of components: 3 
#>        prob      shift       scale     order ar_1       ar_2        
#> Comp_1 0.3729066  5.00821578 0.9543599   2    0.9024640 -0.612078226
#> Comp_2 0.3123368  0.03277879 1.0349174   2   -0.6946305 -0.003218164
#> Comp_3 0.3147566 -5.18855177 4.6730139   2   -0.0829664  0.830994914
#> 
#> Distributions of the error components:
#> 	standard Gaussian
#> 
#> 
#> $Raw[[9]]
#> (To see the internal structure of the object, use function 'str'.)
#> 
#> An object of class "MixARGaussian"
#> Number of components: 3 
#>        prob      shift       scale     order ar_1       ar_2        
#> Comp_1 0.3624051  4.85166715 0.9206978   2    0.8977602 -0.609052782
#> Comp_2 0.2714019 -0.03133673 0.8855050   2   -0.6909813 -0.004544064
#> Comp_3 0.3661930 -4.71720657 5.3220946   2   -0.1426204  0.841405695
#> 
#> Distributions of the error components:
#> 	standard Gaussian
#> 
#> 
#> $Raw[[10]]
#> (To see the internal structure of the object, use function 'str'.)
#> 
#> An object of class "MixARGaussian"
#> Number of components: 3 
#>        prob      shift      scale     order ar_1        ar_2       
#> Comp_1 0.3811939  5.3464447 0.8124193   2    0.90966755 -0.61492823
#> Comp_2 0.2625147 -0.3921913 1.1268423   2   -0.65507046 -0.05019608
#> Comp_3 0.3562914 -3.6015706 4.9267911   2    0.01161045  0.75185519
#> 
#> Distributions of the error components:
#> 	standard Gaussian
#> 
#> 
#> 

mixARExperiment(exampleModels$WL_II, N = 10, raw = TRUE, simargs = list(n = 500))

#> sa$init is:  0 0 
#> Initial vallogf:  -1396.229 
#> Final niter:  11 	vallogf:  -1389.046 
#> Initial vallogf:  -1334.934 
#> Final niter:  11 	vallogf:  -1327.145 
#> Initial vallogf:  -1417.959 
#> Final niter:  11 	vallogf:  -1411.762 
#> Initial vallogf:  -1350.732 
#> Final niter:  11 	vallogf:  -1343.968 
#> Initial vallogf:  -1379.73 
#> Final niter:  11 	vallogf:  -1372.534 
#> Initial vallogf:  -1410.148 
#> Final niter:  11 	vallogf:  -1402.843 
#> Initial vallogf:  -1339.904 
#> Final niter:  11 	vallogf:  -1326.491 
#> Initial vallogf:  -1377.982 
#> Final niter:  11 	vallogf:  -1371.294 
#> Initial vallogf:  -1417.176 
#> Final niter:  11 	vallogf:  -1408.131 
#> Initial vallogf:  -1376.378 
#> Final niter:  11 	vallogf:  -1373.212 
#> [1]  5 23  0  8 62  0
#> [1]  5 23  0  8 62  0
#> [1]  5 23  0  8 62  0
#> [1]  5 23  0  8 62  0
#> [1]  5 23  0  8 62  0
#> [1]  5 23  0  8 62  0
#> [1]  5 23  0  8 62  0
#> [1]  5 23  0  8 62  0
#> [1]  5 23  0  8 62  0
#> [1]  5 23  0  8 62  0

#> $Summary
#> $Summary$mean
#> (To see the internal structure of the object, use function 'str'.)
#> 
#> An object of class "MixARGaussian"
#> Number of components: 3 
#>        prob      shift       scale     order ar_1        ar_2         
#> Comp_1 0.3994682  4.96264835 0.9993621   2    0.89674554 -0.5932192210
#> Comp_2 0.2988788 -0.02267166 0.9768700   2   -0.69941079  0.0002912064
#> Comp_3 0.3016530 -4.94819340 4.9261736   2   -0.02605821  0.8081703054
#> 
#> Distributions of the error components:
#> 	standard Gaussian
#> 
#> 
#> $Summary$sd
#> (To see the internal structure of the object, use function 'str'.)
#> 
#> An object of class "MixARGaussian"
#> Number of components: 3 
#>        prob       shift      scale      order ar_1       ar_2       
#> Comp_1 0.03098521 0.10604208 0.04181184   2   0.01002798 0.008176764
#> Comp_2 0.02333766 0.08751707 0.06768380   2   0.01161332 0.007753148
#> Comp_3 0.02384676 0.43863181 0.39147803   2   0.05072387 0.035812639
#> 
#> Distributions of the error components:
#> 	standard Gaussian
#> 
#> 
#> $Summary$meanUnswitched
#> (To see the internal structure of the object, use function 'str'.)
#> 
#> An object of class "MixARGaussian"
#> Number of components: 3 
#>        prob      shift       scale     order ar_1        ar_2         
#> Comp_1 0.2988788 -0.02267166 0.9768700   2   -0.69941079  0.0002912064
#> Comp_2 0.3016530 -4.94819340 4.9261736   2   -0.02605821  0.8081703054
#> Comp_3 0.3994682  4.96264835 0.9993621   2    0.89674554 -0.5932192210
#> 
#> Distributions of the error components:
#> 	standard Gaussian
#> 
#> 
#> $Summary$sdUnswitched
#> (To see the internal structure of the object, use function 'str'.)
#> 
#> An object of class "MixARGaussian"
#> Number of components: 3 
#>        prob       shift      scale      order ar_1       ar_2       
#> Comp_1 0.02333766 0.08751707 0.06768380   2   0.01161332 0.007753148
#> Comp_2 0.02384676 0.43863181 0.39147803   2   0.05072387 0.035812639
#> Comp_3 0.03098521 0.10604208 0.04181184   2   0.01002798 0.008176764
#> 
#> Distributions of the error components:
#> 	standard Gaussian
#> 
#> 
#> 
#> $Raw
#> $Raw[[1]]
#> (To see the internal structure of the object, use function 'str'.)
#> 
#> An object of class "MixARGaussian"
#> Number of components: 3 
#>        prob      shift       scale     order ar_1        ar_2        
#> Comp_1 0.4072060  4.85467651 1.0437560   2    0.90505966 -0.598065549
#> Comp_2 0.2681891 -0.02621396 0.9303481   2   -0.69029651  0.008237143
#> Comp_3 0.3246049 -4.03958462 5.2190985   2   -0.06985282  0.808585564
#> 
#> Distributions of the error components:
#> 	standard Gaussian
#> 
#> 
#> $Raw[[2]]
#> (To see the internal structure of the object, use function 'str'.)
#> 
#> An object of class "MixARGaussian"
#> Number of components: 3 
#>        prob      shift       scale     order ar_1        ar_2        
#> Comp_1 0.3764546  4.96534512 0.9640374   2    0.87817212 -0.598693832
#> Comp_2 0.3301111  0.05796371 0.8847617   2   -0.70975479  0.005413045
#> Comp_3 0.2934343 -4.48205023 4.7850162   2   -0.06060898  0.792681225
#> 
#> Distributions of the error components:
#> 	standard Gaussian
#> 
#> 
#> $Raw[[3]]
#> (To see the internal structure of the object, use function 'str'.)
#> 
#> An object of class "MixARGaussian"
#> Number of components: 3 
#>        prob      shift      scale     order ar_1         ar_2       
#> Comp_1 0.3580831  5.0538053 1.0459224   2    0.896764150 -0.59725362
#> Comp_2 0.3136001 -0.1038896 0.9580083   2   -0.685405973 -0.00372583
#> Comp_3 0.3283168 -5.3118817 5.0740590   2    0.002895669  0.84250596
#> 
#> Distributions of the error components:
#> 	standard Gaussian
#> 
#> 
#> $Raw[[4]]
#> (To see the internal structure of the object, use function 'str'.)
#> 
#> An object of class "MixARGaussian"
#> Number of components: 3 
#>        prob      shift       scale     order ar_1       ar_2       
#> Comp_1 0.3995995  4.86087160 0.9401739   2    0.8995680 -0.59814705
#> Comp_2 0.3058085 -0.01844111 0.9010362   2   -0.7170398 -0.01150348
#> Comp_3 0.2945921 -4.89342203 4.6705194   2    0.0241223  0.79367672
#> 
#> Distributions of the error components:
#> 	standard Gaussian
#> 
#> 
#> $Raw[[5]]
#> (To see the internal structure of the object, use function 'str'.)
#> 
#> An object of class "MixARGaussian"
#> Number of components: 3 
#>        prob      shift      scale     order ar_1        ar_2        
#> Comp_1 0.3902260  5.1279034 0.9450239   2    0.88616908 -0.601424345
#> Comp_2 0.3123691 -0.1304874 0.9732025   2   -0.69409589 -0.002920161
#> Comp_3 0.2974049 -5.0080883 5.6046092   2   -0.06463516  0.726455906
#> 
#> Distributions of the error components:
#> 	standard Gaussian
#> 
#> 
#> $Raw[[6]]
#> (To see the internal structure of the object, use function 'str'.)
#> 
#> An object of class "MixARGaussian"
#> Number of components: 3 
#>        prob      shift       scale     order ar_1       ar_2        
#> Comp_1 0.3787557  4.97342212 0.9744767   2    0.9017796 -0.583605610
#> Comp_2 0.2866156  0.06813088 0.9918097   2   -0.7021454  0.006129506
#> Comp_3 0.3346287 -5.47185953 4.5670192   2   -0.0658921  0.851774199
#> 
#> Distributions of the error components:
#> 	standard Gaussian
#> 
#> 
#> $Raw[[7]]
#> (To see the internal structure of the object, use function 'str'.)
#> 
#> An object of class "MixARGaussian"
#> Number of components: 3 
#>        prob      shift      scale     order ar_1        ar_2        
#> Comp_1 0.4729701  4.8833735 1.0135978   2    0.89529576 -0.584369615
#> Comp_2 0.2697326  0.1052054 0.9647773   2   -0.71299841 -0.001341753
#> Comp_3 0.2572973 -5.4347411 4.2541827   2    0.07336466  0.803745665
#> 
#> Distributions of the error components:
#> 	standard Gaussian
#> 
#> 
#> $Raw[[8]]
#> (To see the internal structure of the object, use function 'str'.)
#> 
#> An object of class "MixARGaussian"
#> Number of components: 3 
#>        prob      shift       scale    order ar_1        ar_2       
#> Comp_1 0.4091667  4.84697392 1.002041   2    0.91286510 -0.59781174
#> Comp_2 0.3051565 -0.09821069 1.101045   2   -0.70381605  0.01311129
#> Comp_3 0.2856768 -4.80044631 4.792410   2   -0.06191883  0.81903139
#> 
#> Distributions of the error components:
#> 	standard Gaussian
#> 
#> 
#> $Raw[[9]]
#> (To see the internal structure of the object, use function 'str'.)
#> 
#> An object of class "MixARGaussian"
#> Number of components: 3 
#>        prob      shift       scale     order ar_1        ar_2        
#> Comp_1 0.4138008  4.94434420 1.0534299   2    0.88947325 -0.577583042
#> Comp_2 0.2712318  0.04100286 0.9956208   2   -0.68328932 -0.008685234
#> Comp_3 0.3149675 -4.96166954 5.2692977   2   -0.05325865  0.842065524
#> 
#> Distributions of the error components:
#> 	standard Gaussian
#> 
#> 
#> $Raw[[10]]
#> (To see the internal structure of the object, use function 'str'.)
#> 
#> An object of class "MixARGaussian"
#> Number of components: 3 
#>        prob      shift      scale    order ar_1        ar_2        
#> Comp_1 0.3884196  5.1157679 1.011162   2    0.90230866 -0.595237798
#> Comp_2 0.3259735 -0.1217767 1.068091   2   -0.69526585 -0.001802461
#> Comp_3 0.2856069 -5.0781906 5.025524   2    0.01520178  0.801180894
#> 
#> Distributions of the error components:
#> 	standard Gaussian
#> 
#> 
#> 

## the following take more time.
##
## mixARExperiment(exampleModels$WL_II, N = 1000, raw = TRUE, simargs = list(n = 500))
## mixARExperiment(exampleModels$WL_II, N = 1000, raw = TRUE, simargs = list(n = 50))
## mixARExperiment(exampleModels$WL_II, N = 1000, raw = TRUE, simargs = list(n = 50))
## mixARExperiment(exampleModels$WL_II, N = 1000, raw = TRUE, simargs = list(n = 50),
##     estargs = list(maxniter = 500))
## mixARExperiment(exampleModels$WL_II, N = 1000, raw = TRUE, simargs = list(n = 50),
##     estargs = list(init = 1, maxniter = 500))
## mixARExperiment(exampleModels$WL_II, N = 1000, raw = TRUE, simargs = list(n = 100),
##     estargs = list(init = 1, maxniter = 500))
## mixARExperiment(exampleModels$WL_II, N = 1000, raw = TRUE, simargs = list(n = 100),
##     estargs = list(init = 5, maxniter = 500))
## mixARExperiment(exampleModels$WL_II, N = 1000, raw = TRUE, simargs = list(n = 100),
##     estargs = list(init = 10, maxniter = 500))