Class "mixComp" - manipulation of mixAR time series

Class "mixComp" represents components of mixture autoregressive time series and their transformations obtained by arithmetic and related operations. Methods are provided to allow convenient computation with such time series.

Objects from the Class

Objects can be created by calls of the form new("mixComp", ...). It is more usual however to obtain such objects initially from functions such as mix_ek. Methods are defined to allow for convenient and intuitive further manipulation of such objects.

Internally, an object of class mixComp is a matrix with one column for each component. However, methods for arithmetic operations involving mixComp objects are defined to perform natural operations for mixture objects. For example, multiplication by vectors is commutative and ``does the right thing''.

Slots

m:

Object of class "matrix" with one column correponding to each component of the mixture AR model.

Methods

Arithmetic operations involving mixComp objects are defined to allow for convenient execution of computations for mixture autoregressive models, see mixComp.

-

signature(e1 = "mixComp", e2 = "missing"): unary minus for "mixComp" objects.

-

signature(e1 = "numeric", e2 = "mixComp"): If e2 is thought of as a matrix, \(m\), then the number of elements of e1 must be the same as the number of rows of \(m\) and each column of \(m\) is subtracted from e1, see also "mix_ek", "mix_hatk". As a special case, if \(m\) has only one row, then it is subtracted from each element of e1, i.e. that row is replicated to obtain a matrix with as many rows as the length of e1 and then subtracted from e1 as above. The result is a mixComp object.

-

signature(e1 = "mixComp", e2 = "numeric"): This is analogous to the above method. (FIXME: the code of this function does not deal with the special case as in the above method. Is this an omission or I have done it on purpose?)

%of%

signature(e1 = "function", e2 = "mixComp"): This applies the function e1 to each element of e2. Together with the arithmetic operations this allows for easy computation with mixComp objects (e.g. pdfs, likelihoods).

%of%

signature(e1 = "character", e2 = "mixComp"):

%of%

signature(e1 = "list", e2 = "mixComp"): If e1 is of length one it specifies a function to be applied to each element of e2, otherwise it is a list of functions, such that the \(i\)th function is applied to the \(i\)th column of e2@m.

*

signature(e1 = "mixComp", e2 = "mixComp"): ...

*

signature(e1 = "mixComp", e2 = "numeric"): see the following.

*

signature(e1 = "numeric", e2 = "mixComp"): ``Column'' \(i\) of the mixComp object is multiplied by the \(i\)th element of the numeric vector, i.e. each ``row'' of the mixComp object is multiplied by the vector (or, the vector is replicated to a matrix to be multiplied by the mixComp object).

*

signature(e1 = "function", e2 = "mixComp"): Multiplying a function by a mixComp object actually applies the function to each element of the object. TODO: This is a misuse of methods, see operator %of% which does the same.

*

signature(e1 = "character", e2 = "mixComp"): The first argument is a name of a function which is applied to each element of the mixComp object. This is a misuse of methods, see operator %of% which does the same. (TODO: but the latter does not have a method with this signature.)

/

signature(e1 = "mixComp", e2 = "numeric"):

/

signature(e1 = "numeric", e2 = "mixComp"): Division works analogously to "*".

^

signature(e1 = "mixComp", e2 = "numeric"): If k is a scalar, raise each element of e1@m to power k. TODO: for consistency this operation should have the semantics of "*" and "/" but this operator probably makes sense only for scalar 'e2', where the semantics doesn't matter. So, don't bother for now.

+

signature(e1 = "numeric", e2 = "mixComp"):

+

signature(e1 = "mixComp", e2 = "numeric"): Addition involving mixComp objects works analogously to subtraction.

inner

signature(x = "mixComp", y = "missing", star = "missing", plus = "missing"): With one argument inner computes the sum of the columns of the argument. This is conceptually equivalent to y being a vector of ones.

inner

signature(x = "mixComp", y = "numeric", star = "missing", plus = "missing"):

inner

signature(x = "numeric", y = "mixComp", star = "missing", plus = "missing"): The number of elements of the numeric argument should be equal to the number of rows of the mixComp object. Effectively, computes the inner product of the two arguments. The order of the arguments does not matter. (TODO: FIXME: make sure that this is true!) Returns a numeric vector.

inner

signature(x = "mixComp", y = "numeric", star = "ANY", plus = "ANY"): Computes a generalised inner product of x with y using the specified functions in place of the usual "*" and "+" operations. The defaults for star and + are equivalent to multiplication and addition, respectively. Note that "+" is a binary operation (not \(n\)-ary) in R. So technically the correct way to specify the default operation here is "sum" or sum. Since it is easy to make this mistake, if plus == "+", it is replaced by "sum". (In fact, plus is given a single argument, the vector of values to work on. Since "+" works as a unary operator on one argument, it would give surprising results if left as is.)

inner

signature(x = "mixComp", y = "numeric", star = "ANY", plus = "missing"): This is a more efficient implementation for the case when plus = sum.

mix_ncomp

signature(x = "mixComp"): Number of components.

signature(x = "mixComp")

A "mixComp" object is essentially a matrix. This method gives the dimension of the underlying matrix. This method indirectly ensures that nrow() and ncol() work naturally for "mixComp" objects.

Examples

## dim, nrow, ncol
a <- new("mixComp", m = matrix(c(1:7, 11:17, 21:27), ncol = 3))
a

#> An object of class "mixComp"
#> Slot "m":
#>      [,1] [,2] [,3]
#> [1,]    1   11   21
#> [2,]    2   12   22
#> [3,]    3   13   23
#> [4,]    4   14   24
#> [5,]    5   15   25
#> [6,]    6   16   26
#> [7,]    7   17   27
#> 

dim(a)

#> [1] 7 3

nrow(a)

#> [1] 7

ncol(a)

#> [1] 3

mix_ncomp(a)

#> [1] 3

-a

#> An object of class "mixComp"
#> Slot "m":
#>      [,1] [,2] [,3]
#> [1,]   -1  -11  -21
#> [2,]   -2  -12  -22
#> [3,]   -3  -13  -23
#> [4,]   -4  -14  -24
#> [5,]   -5  -15  -25
#> [6,]   -6  -16  -26
#> [7,]   -7  -17  -27
#> 

a - 1:7

#> An object of class "mixComp"
#> Slot "m":
#>      [,1] [,2] [,3]
#> [1,]    0   10   20
#> [2,]    0   10   20
#> [3,]    0   10   20
#> [4,]    0   10   20
#> [5,]    0   10   20
#> [6,]    0   10   20
#> [7,]    0   10   20
#> 

1:7 + a

#> An object of class "mixComp"
#> Slot "m":
#>      [,1] [,2] [,3]
#> [1,]    2   12   22
#> [2,]    4   14   24
#> [3,]    6   16   26
#> [4,]    8   18   28
#> [5,]   10   20   30
#> [6,]   12   22   32
#> [7,]   14   24   34
#> 

2*a

#> An object of class "mixComp"
#> Slot "m":
#>      [,1] [,2] [,3]
#> [1,]    2   22   42
#> [2,]    4   24   44
#> [3,]    6   26   46
#> [4,]    8   28   48
#> [5,]   10   30   50
#> [6,]   12   32   52
#> [7,]   14   34   54
#> 

b <- new("mixComp", m = matrix(rnorm(18), ncol = 3))

## apply a function to the columns of a mixComp object
pnorm %of% b

#> An object of class "mixComp"
#> Slot "m":
#>           [,1]      [,2]      [,3]
#> [1,] 0.1040102 0.4427310 0.8958118
#> [2,] 0.6015682 0.1650989 0.9322777
#> [3,] 0.2048865 0.5223070 0.6453924
#> [4,] 0.1444304 0.6887628 0.5625665
#> [5,] 0.7659290 0.6923580 0.5308246
#> [6,] 0.2394938 0.1411296 0.6016213
#> 

## apply a separate function to to each column
flist <- list(function(x) pnorm(x),
              function(x) pt(x, df = 5),
              function(x) pt(x, df = 4) )
flist %of% b

#> An object of class "mixComp"
#> Slot "m":
#>           [,1]      [,2]      [,3]
#> [1,] 0.1040102 0.4455439 0.8615967
#> [2,] 0.6015682 0.1874590 0.8951327
#> [3,] 0.2048865 0.5212236 0.6359326
#> [4,] 0.1444304 0.6783354 0.5587517
#> [5,] 0.7659290 0.6816846 0.5289675
#> [6,] 0.2394938 0.1656925 0.5952680
#>