Simulate from multivariate mixAR models

Simulate data from multivariate mixAR models under the assumptions of multivariate Gaussian innovarion

mixVAR_sim(model, n, init, nskip = 100, flag = FALSE)

Arguments

model
model from which to simulate, an object inheriting from class mixVAR.

n
size of simulated multivariate series.

init
initial values, a numeric matrix. If missing, a matrix of 0 values is generated.

nskip
number of burn-in values.

flag
if TRUE returns also the regimes.

Details

mixVAR_sim simulates a series of length nskip + n and returns the last n values. init provides initial values for the algorithm. Each row is considered as a time point. The number of rows must be at least equal to the maximal AR order.

Value

a numeric matrix with n rows.

See also

Examples

AR <- list()
AR[[1]] <- array(c(0.5,-0.3,-0.6,0,0,0.5,0.4,0.5,-0.3), dim = c(3,3,1))
AR[[2]] <- array(c(-0.5,0.3,0,1,0,-0.5,-0.4,-0.2, 0.5), dim = c(3,3,1))

prob <- c(0.75, 0.25)
shift <- cbind(c(0,0,0), c(0,0,0))

Sigma1 <- cbind(c(1, 0.5, -0.4), c(0.5, 2, 0.8), c(-0.4, 0.8, 4))
Sigma2 <- cbind(c(1,0.2, 0), c(0.2, 2, -0.15), c(0, -0.15, 4))
Sigma <- array(c(Sigma1, Sigma2), dim = c(3,3,2))

m <- new("mixVARGaussian", prob=prob, vcov=Sigma, arcoef=AR, shift=shift)
mixVAR_sim(m, n=500, init=matrix(rep(0,3), ncol=3), nskip=100, flag=FALSE)

#>          [,1]         [,2]         [,3]
#>  -3.129153542 -0.604809322  2.977197198
#>   2.360562827 -3.394232690  1.082076194
#>  -5.000190759  1.330780938  0.102966677
#>  -2.433116170  0.513043133  3.407589475
#>  -1.468105945  0.634979358 -0.793292726
#>  -0.261572261  1.509040448 -0.095409298
#>  -0.996408876 -0.344230080  0.354208216
#>  -0.376591634 -0.453245687 -0.819532183
#>  -1.049573975 -0.627953727  0.349239923
#>   0.216110995  1.024446575  1.016363350
#>  -0.434798324  0.171570461  0.506778371
#>  -1.271801969 -1.053926095 -0.782668025
#>  -1.651060441 -2.758319093  2.544524448
#>   0.565062728  2.788280723 -0.102177062
#>  -0.144527824 -3.639308339  0.511597533
#>  -3.971922730 -2.739329017  3.731639219
#>  -1.742549234  1.842073107 -3.114960770
#>  -2.364031155  0.242096043  3.155577265
#>  -0.146559844  4.363174981  3.740881074
#>   0.870679216  0.942953035  1.976211350
#>   0.287402692  1.633403616 -1.735522183
#>  -0.326376705 -0.528167594 -0.354693795
#>  -1.675500141 -2.805868041  0.075598875
#>  -0.937668565 -0.105184146 -0.677044828
#>  -2.059569087  0.712598576  1.480261884
#>  -1.237626330  0.845515304  2.710859462
#>   0.435110333  4.793334101  2.953185524
#>   2.564802013 -2.065609510  0.417696695
#>   1.754248771  0.119112583 -4.454644770
#>   1.112127177  1.290851490 -0.326279568
#>   0.636333854 -0.988654680 -0.662541207
#>   0.056494477 -1.278655472 -0.400339937
#>  -1.715988584 -0.885276269  1.387637784
#>  -0.548630691  0.911316510  1.116519414
#>   1.282396538  1.251127761 -0.738547575
#>   1.538077636  0.466122473 -1.754086619
#>   0.620093918  0.807480701  0.500694895
#>   1.449868997  2.130287034  0.507162554
#>  -0.512226612 -0.111722927  0.842257392
#>  -1.228691503 -0.838000283 -0.399387319
#>  -2.678281112  0.490699369 -0.952171314
#>   0.806534621 -1.715048180 -1.755224850
#>  -1.262883696 -1.540344623 -2.188093201
#>  -2.185500493 -0.525470963  2.818463355
#>  -0.576964633  2.652678170  2.605242167
#>  -2.078895526  0.726165522  2.864684685
#>   0.130838799  2.995160154  0.594214372
#>   1.218025983  0.588354762  0.440939796
#>   0.556445573 -0.388033548  2.063800557
#>  -0.217895578  0.548461250  0.450950996
#>  -0.154635266 -1.413237274 -1.246212543
#>  -1.052507817  0.402768354  4.014495446
#>   1.481839443  1.888879127  1.585767037
#>  -0.936060810 -1.362651090 -2.454491451
#>  -2.116843836 -3.113799486  2.983579109
#>  -2.428994562  0.375193759  2.928757593
#>   0.888269697  3.143609901  2.695103165
#>   0.636627928  0.174178367  1.160767353
#>  -1.538259584  2.575979496  0.877276104
#>   0.043229606  1.904291527  5.357303601
#>   1.635075526  2.410276421 -0.368403510
#>   0.579996978 -0.726485920 -0.507222023
#>   0.899542852 -0.469687452 -0.839909831
#>  -0.139024814 -1.832737925  0.200284062
#>  -0.316496158  0.378422184 -0.593177739
#>   0.005157011  0.525716436  1.193148164
#>   1.206685352 -0.178332337 -1.225797386
#>   0.045146623 -0.730150132  1.862664517
#>   0.686161569  0.539363082 -2.690740753
#>   2.351761041  1.367299882 -0.911060902
#>   0.013967129 -0.157387961 -2.363644748
#>   0.355224734 -0.390627867  4.447585569
#>   2.248061673  2.682518352 -3.132419852
#>   0.823934174 -2.348516311  1.668476966
#>  -0.081885972  1.273747004 -0.948101255
#>  -2.636486960 -2.346067330 -0.907630075
#>  -0.556788835  1.117423998 -0.918980984
#>  -0.419636814  0.330092399 -1.777049459
#>  -0.893390494 -0.262296693  1.861146092
#>  -0.950334207  0.154771047  0.724045337
#>   0.025457118  2.065218082  2.533755170
#>   1.554746389  3.791632417  0.985667249
#>   1.797681128  0.417869089 -1.536765242
#>   0.313923353 -1.665453267 -0.115739886
#>  -1.827296235 -0.407701940 -1.549128051
#>  -2.997742264 -3.294530237  1.528671358
#>  -0.932720581  2.577258819  0.427679155
#>  -0.864064586 -0.784063949 -0.894848631
#>  -1.072258912  3.158454478  2.455232648
#>  -1.158091932 -1.926204428 -0.083080731
#>   0.017718608  0.654224124 -1.478406783
#>  -1.258181690 -0.381495301  2.012736878
#>   2.970235595  1.713161003 -1.095756873
#>   1.150128160 -2.004441850  1.003665718
#>   0.070582113 -0.432499489 -5.589101310
#>   2.315893476  0.802331139  0.506593430
#>  -0.151549276 -2.015464785  1.354298122
#>  -2.506591636 -2.816816419  3.169164604
#>  -3.622333869 -0.932797218  5.071373548
#>   0.079330646  4.841096349  3.520977933
#>   2.741630835 -2.002500064  0.609651335
#>  -3.417310452  1.004997673  2.156076843
#>   2.105951753 -0.636273604 -1.900726549
#>   1.167885250  1.224484344 -0.603787077
#>   0.511761727  1.132694578  1.633787049
#>   1.062121677  1.269380247 -0.992608184
#>  -0.812715376 -1.284303147  2.914171576
#>   1.952913778  0.550640743 -1.183617965
#>   0.561726156 -2.958055718 -0.466371047
#>  -0.838887321 -2.790125367 -0.628184871
#>  -2.766045886 -2.381198476  5.400056670
#>   1.795029127  5.359645241 -0.687655567
#>   2.290247914 -0.367342184 -5.933018110
#>   0.404149346  2.199950226 -1.737965244
#>   2.644677285 -1.565117679 -0.747854301
#>  -3.523031399 -2.031756434 -2.291263916
#>  -1.728526367  0.295119831  0.173880380
#>  -4.331856845  0.708953238  2.986857053
#>  -1.100431554  1.359618541 -0.919393536
#>  -1.410949076 -1.007895237 -0.236878740
#>  -1.173941604  1.065101494 -0.715805594
#>  -0.551152351  0.539282125  2.331354811
#>  -0.559809811  0.130116401  1.157346191
#>   1.482177872  1.044942714 -1.493981780
#>   1.613574226 -1.146789209  0.125910960
#>   2.377502492  0.555427523 -0.389571545
#>  -2.994290241 -1.099141507 -2.436021151
#>  -1.850728244  0.261660168 -0.667572498
#>   0.221008266  5.406796796  4.206444040
#>   0.751061400  0.025984229 -0.110111559
#>   1.886658257 -0.151521651 -1.319939501
#>   1.961199862  1.192838160 -2.054803315
#>   0.454460390 -2.630664562 -2.122244922
#>   0.202614477 -1.254311562 -0.003054531
#>  -2.322350941  0.432029498 -3.637097100
#>   2.011888422  1.394891881 -3.419322993
#>  -0.437292196 -4.340908091 -2.996836407
#>  -1.354533835 -0.959411547 -2.780571819
#>  -1.268935650  1.114565074  1.110628003
#>  -0.327816740  0.833695344  1.364162985
#>   0.716339421  0.216225093  1.375330663
#>   0.084273315  0.681852666  0.860180929
#>  -0.682112312  0.628168819  1.006726708
#>  -0.129914158 -0.170278292 -1.068785159
#>   0.332214336 -1.768540999  0.850683553
#>   0.244997162 -1.574458150 -3.325050706
#>  -1.960241843 -1.401479133 -0.223904703
#>   0.163096893  0.605590924 -0.652718288
#>  -0.890959652 -1.722315240  0.143141452
#>  -1.271254200  0.173374574  0.558114035
#>  -2.283321686 -0.279601970  1.466628723
#>   0.146279098 -2.706759680  1.732447215
#>   0.924236856  2.387702911 -0.936452801
#>  -0.526092737 -0.091477011  0.475163475
#>   0.663177738  1.811282247 -0.401729228
#>   0.864490067 -3.258871774 -5.189497961
#>  -2.934324232 -3.746512788 -5.386218431
#>  -2.562220853  3.098000424 -0.272168102
#>  -1.345308335  0.933015437  5.657193879
#>  -0.832025594 -0.492783905  2.535846878
#>  -1.748891237 -1.977789363 -1.993476895
#>  -2.034794429 -2.780988819 -0.668171756
#>  -2.598565453  0.462895302  2.801013420
#>   0.883823593  0.682402795 -3.623616123
#>  -1.374923916 -3.780499997 -0.432011376
#>  -2.606720999 -0.519366743  6.212264282
#>   1.608826744  5.105116563 -2.312798187
#>  -0.255573425 -0.434516776  3.852137008
#>  -0.018245378  0.845237949 -1.178581241
#>  -0.579963940 -1.669155002 -1.699038987
#>   0.465240724  1.004210583  2.168335610
#>   0.030883947  1.223259764  1.388703141
#>   0.500065673 -0.597007416 -2.435382397
#>  -0.166689252 -0.007751806 -0.158279662
#>  -0.657042836  0.646382735 -1.619084747
#>  -1.325831945 -0.529938469  3.010303385
#>  -0.857096987  2.046424098 -1.886524436
#>  -1.504900403 -0.747546994  2.144573013
#>  -0.311469168  0.233375964  0.366867956
#>   0.092275856 -0.176279531  1.168101875
#>   1.892939188  1.561001087  0.966176867
#>   1.884940714  1.269504019  0.410683985
#>   0.822065681  0.236155526  0.961530962
#>  -0.134581090  0.650004036  1.884235816
#>   0.082551784  0.778404839 -0.325964375
#>   0.965984539 -0.995240459  0.276701111
#>  -0.143105768 -0.542181412  3.023985382
#>   1.623318846  3.237446812  0.188222287
#>   2.240204536  2.647704754  1.076601702
#>   2.129213376  2.348060053  1.853024747
#>   3.345093096  0.902575915 -0.137278530
#>   1.957059080  0.701498207  0.661216968
#>  -0.191825339  1.097317173  1.912019028
#>  -1.014980549  1.320477645 -1.505480042
#>  -0.529030829  2.053062959  2.581654771
#>   0.031732483  0.714970647  0.467720936
#>  -0.762533334 -0.063827041  0.187880890
#>  -0.415543972  0.787005629  0.360736643
#>  -0.538662971 -0.595636816 -0.609934543
#>  -1.062752969 -1.154014150 -1.184371181
#>  -1.016128155  0.214779516  0.106269029
#>   0.487805084 -0.520998087  1.577520801
#>   0.624674247  0.509752977 -4.497362391
#>  -2.115283150 -0.660954046  3.428360802
#>   0.444156340  2.110327158 -1.162687048
#>  -0.401080752 -1.849142102 -0.210804562
#>   1.185931522  2.197893556  4.504468452
#>   3.167558330  1.938123482 -0.430932333
#>   0.867702549  2.863185088 -1.910581051
#>   1.430163713 -1.783813241  0.689316560
#>   0.542685977  0.029347916  1.525456534
#>   0.645863822  0.822842057 -3.791669970
#>  -1.894866308 -1.468665799 -0.485495905
#>  -1.262014149 -0.977669216 -0.414219052
#>  -1.812926745 -0.117011443  0.738304717
#>  -0.047605680 -3.487022373 -0.343206738
#>  -5.448240681 -1.233426863  2.423451423
#>  -2.704096133  1.281394638  3.222756855
#>  -0.187388305 -3.002977197  1.696361227
#>  -1.120802267  1.354298074 -0.889781123
#>   3.093428459  1.378230635 -3.767083752
#>  -0.704013842 -5.123557001  1.746812336
#>   2.129845848  3.275939477 -2.621635350
#>  -0.366195619 -2.099047512  3.083781098
#>  -3.836312095 -2.003185882  3.487436133
#>  -1.494218003  2.639201194  1.726457408
#>   4.434658783 -1.947299973 -2.208164548
#>   1.097717355  0.224290825 -1.200960012
#>   0.481252272 -2.321539333 -3.121334657
#>   0.068108304  0.934433437 -1.306168710
#>  -1.079798081  1.266130059  4.879421825
#>  -0.742815055  2.266775988  0.246340907
#>   3.201063237  0.574292352 -1.765891494
#>   2.177761904 -3.503903081 -1.936953946
#>  -4.192931941 -2.070803365  7.964067117
#>   2.415044892  6.939527075 -0.689945205
#>   1.490331318 -2.287892417 -0.604571158
#>  -1.414096050  1.651787111  1.094540501
#>   2.013302487 -1.075253331  1.273535261
#>  -2.476522700 -0.035055251  2.348311797
#>  -0.775943049  2.303691496  2.242196776
#>   0.345768633 -0.766655569  0.384258396
#>  -1.144782740 -0.603197320  1.231610907
#>   0.902910312  2.544512812 -0.441941776
#>  -0.442588960 -1.694838878  2.557171451
#>   0.750667844  0.532780253  0.046157220
#>  -0.454941981  0.677461493  2.989958987
#>   1.156730913  2.705898186 -1.075830495
#>   1.201632990 -0.564087538 -0.805101294
#>   0.468837864  0.563634426 -1.972059380
#>  -1.004410494 -2.140806495  1.176708077
#>   0.498038893  0.995214980 -1.594684530
#>  -0.524465799 -0.076360480  1.013654071
#>  -0.205186698  3.319526646  1.817468915
#>   2.587038084  1.660313154  1.767411540
#>   1.679074724  1.169364585  1.207997052
#>   0.572924496  0.317946153 -0.689445707
#>  -0.423032084  2.014118490  0.870407740
#>  -0.272321275  1.745091809  3.696527793
#>   2.855677586  2.940787590 -5.781321296
#>  -0.723531515 -4.224690915 -0.036347509
#>   0.564704543 -0.375746604 -3.441224577
#>  -1.261332153 -1.285685271  3.809764969
#>  -2.878254588  0.550950336  3.171078970
#>  -0.750048510  3.128277785  1.026472292
#>   3.867989219  1.333316931  1.024038622
#>  -0.140656134 -0.409108063 -3.266925759
#>  -1.147336479 -3.288602432 -0.081097259
#>   0.439048584  2.939094120  0.310801185
#>   2.471343296  1.597167351 -4.365638615
#>  -1.268745980 -4.286472567  0.711164129
#>  -4.553350705 -1.969569854  0.670645172
#>  -0.429457044 -1.815889399  2.950963022
#>   0.625567273  2.984555902  1.691217592
#>   0.972479621  1.471413912  1.009343756
#>   1.220997430  2.188179368  4.371749492
#>   2.356364986  2.869571847 -2.955725104
#>  -0.564267892 -3.541931980  3.161346544
#>  -5.015166176 -1.449069477  2.633629969
#>  -0.935368082  0.509827252  2.345400023
#>  -0.406887853  0.898112881 -0.197893256
#>   0.595548011  0.143750846  2.343134915
#>   1.166251577  0.749672853 -2.094505521
#>   1.563747635  1.993273852 -1.261245906
#>   1.425454262 -0.560026592 -5.376093307
#>  -2.220905215 -4.529988959  2.485981287
#>  -1.217688644  3.553616672  0.814462277
#>  -1.142134411 -1.582383794  0.823001614
#>  -0.300487994 -0.672162855 -0.917683905
#>  -1.257767579  1.749695706  1.171089126
#>   0.036673676  2.287599398  4.827620826
#>   1.911493379  3.100148248  0.567631599
#>   1.194826860 -0.672618541 -0.272115008
#>   0.487423801 -1.065805760 -2.705648598
#>   0.343202029  1.411863427 -2.131657556
#>  -1.246537814 -0.638008647  2.655782175
#>   1.432119968  3.651088691  3.336044809
#>   0.661852093  2.040975406  4.450206720
#>   3.729935221  1.831522959 -0.642725118
#>   2.109458042 -1.539508506 -3.000691321
#>  -1.715782022  2.754471402  3.299553654
#>   0.937468318 -0.241937818 -5.026212006
#>  -1.414840028 -3.550965449 -0.925655348
#>  -2.315420416  1.217610201 -1.158612621
#>  -2.412792248  1.362288209  4.181586771
#>   0.784727097  2.531069588  0.262304586
#>   0.814017444 -0.729426872 -0.505530345
#>   1.527685554 -1.871287269 -3.179538389
#>  -2.451002093  0.159839143  0.513209176
#>  -1.413282886 -0.089817742  0.604232469
#>  -0.338728451  0.349509146  2.612432700
#>   0.263039385  0.937116736  0.240172256
#>  -0.835116137 -1.115828807  1.094714658
#>  -2.112092212 -1.185973188  1.411426011
#>  -0.571212762  0.594168708  0.745296795
#>   1.471478204  2.024000748  0.933454637
#>   1.576622974 -0.227211821 -2.162481504
#>   0.763242310 -1.924038334 -3.328581452
#>  -1.727393348 -2.381439527  0.995816887
#>   0.275166015 -0.052298926  0.287876239
#>   0.442488778 -0.998595918 -1.924938822
#>  -1.591696225 -0.050359606  1.159466738
#>  -1.250840689  0.066492912 -1.391953220
#>  -0.702556204  1.903210142  1.513863412
#>   2.743913201  2.111300449  4.291368906
#>  -2.331618744 -0.973518799  1.964558862
#>  -0.098728456  2.102291689  2.822005191
#>   0.371914857  1.332521624  1.581602142
#>   2.094038674  3.319332542  0.137104475
#>   0.624062951 -0.621694773 -1.040021881
#>   0.373949469 -1.169130331  4.282881334
#>   2.807401800  1.470082220 -4.045548533
#>   1.314670261 -1.367835639 -3.971891322
#>   0.943200584 -2.945210704 -2.874146540
#>  -0.874722940 -1.912196184  1.263067098
#>   0.432138935  0.293646753  0.361641629
#>  -0.328074425 -0.033372278  0.270398367
#>  -0.644296437 -0.828459444  3.219585257
#>   0.302068744  0.065740814 -3.898418191
#>  -1.099851967 -3.285011686 -0.188132110
#>  -0.548460200 -0.311938820 -1.487841157
#>  -0.874468445  0.985422899  4.695126866
#>  -1.341069732  0.202910834  3.470927999
#>  -2.139419270  1.528157314  1.368900054
#>   0.778610417 -0.265545908  0.620467056
#>  -1.448652773  0.563707628  2.059379967
#>   0.505202973  2.232868128 -1.128356292
#>   0.813976735 -0.474102051  1.356978251
#>   1.046606251  1.208134918 -0.082032768
#>   1.237673591  1.342541734  3.155778653
#>   2.875723691  0.863354173 -4.920089276
#>   1.372143222 -0.340717876 -0.697548868
#>  -0.878689693 -1.458319866  1.135173395
#>  -0.577008850  1.086030385 -2.691488953
#>  -2.628625090 -0.413615369  3.112171842
#>  -0.244634520  1.241400773  0.887517168
#>   1.160835684  0.757720888  1.828566104
#>   2.085670702  0.782625764 -0.567679603
#>   0.301680320 -1.996465536 -1.352541405
#>  -2.559047383 -0.650389037  2.013034967
#>   0.372456718  3.901951802  0.866990700
#>  -1.375846135 -0.974903864  0.517681253
#>   0.333423395  2.170094299  3.522090132
#>   1.539022599  2.639809398 -0.596042598
#>   1.574240545  1.148497136 -3.307821875
#>   0.596489910 -1.426695373  0.261170106
#>  -0.145947331 -2.221017014 -6.192841491
#>   0.875373506  0.782586537 -3.187308763
#>   0.664337025  1.581908444  1.192337100
#>   0.942333973  2.028551895  3.418774718
#>   2.034361329  1.411967534 -0.439056582
#>   0.778531022  1.245604459 -2.359074308
#>  -1.164521105 -3.889214685  0.052836664
#>  -0.855152549  1.363584053 -1.078794729
#>  -1.839872916 -1.317738756  0.860495065
#>  -1.425723752  2.151406278  1.073073272
#>  -0.027236569 -1.398909569 -1.279760907
#>  -0.701057158  0.874407982 -0.636198588
#>   0.638401449  0.797478089  0.640944907
#>   0.119736108  1.979431762  1.457674235
#>   0.750568849  0.736545210 -0.705865294
#>   1.879016447  0.885134643  1.779381829
#>   0.029588023 -1.545525596  2.440761333
#>   0.943633819  0.460042652 -1.809643729
#>   0.319148256  2.023904158  0.237170344
#>   2.191815264 -1.382620985 -0.611860017
#>   1.867427346  1.176180726 -3.499158539
#>  -0.292388710 -3.238597482 -2.696980447
#>  -3.183636504 -1.215575355 -3.267333585
#>  -2.748944436 -0.218588603  3.730941103
#>   0.940725965  0.208527533 -1.084239382
#>  -1.145806961 -0.620600923  2.022487214
#>   0.635518000  1.239296643 -1.021877229
#>   0.009664137  0.707297941  2.547621172
#>   0.168889762  1.099942738  0.972591243
#>   2.006940927 -1.447517454  0.529257803
#>  -4.434315678  1.110326075 -3.783147760
#>  -4.064775270 -0.393138835  5.414094438
#>  -0.575802315  3.027770423  0.665603308
#>  -0.545262216  1.163218763  2.943510905
#>  -0.988841750 -1.364825277  3.382496267
#>   0.071982010  1.252864317  0.735748351
#>  -0.856262852 -0.659888127 -0.844372466
#>  -2.303401195 -0.817133773  1.017836813
#>  -1.481778322  1.018201208  1.623288190
#>   0.111852878  0.340872208 -0.889576890
#>   0.318033004  0.857231871  0.189789145
#>  -0.613852320 -3.285291875 -2.397405362
#>  -1.675623498  0.531156316  0.210288658
#>  -0.318524489  0.552913704 -0.361882855
#>  -0.119636664 -0.961621199  1.618071809
#>  -0.376101859  2.089292042  2.539666693
#>  -0.739585743 -0.137739407  2.272469074
#>   2.296137439  0.806538570 -2.622087640
#>   0.968011982 -0.187727085 -1.057801671
#>  -0.037791923 -1.155544506  1.082362173
#>  -1.431353446  0.979737734  1.100291592
#>   0.266803959  2.087727880 -0.217314855
#>   0.203345602 -2.239664515  1.443596477
#>  -2.097989136 -0.178299337  2.979352286
#>   0.369843609  0.421259250  0.133061749
#>   1.097222167  0.999023449 -0.553915948
#>  -0.120267171 -1.987635118 -2.762384918
#>  -0.973765667  1.154815195  4.180912092
#>  -0.905212855 -3.258450349  6.030823322
#>   2.933159132  3.642982079 -1.672655666
#>   1.308911357 -0.998074792  2.880768363
#>   1.634130938  1.977605375  0.620296492
#>  -0.972717292 -2.513682072  2.231643007
#>   1.464919519  1.997491885 -1.893752751
#>   0.341318625 -3.603613356 -0.670198395
#>  -0.280851824  2.128165361 -3.302375953
#>   2.694518374 -0.928004741 -6.213939038
#>  -0.840293116 -2.742472972 -0.302213814
#>   1.089841367 -1.808328485 -1.214206593
#>  -1.898548674  2.700402034  0.492695507
#>  -1.635236101  0.957541898  2.322940883
#>  -0.080327053 -0.058623920  1.613138341
#>   2.060599227  1.148342574  1.379251444
#>   3.255810307  1.527427878 -2.454471171
#>  -1.462951012 -1.542524688 -0.450774287
#>  -1.020259653  0.012812015  0.284784037
#>  -1.519435359  0.180884991  0.095759941
#>  -0.487715561 -2.346716801 -2.088651619
#>  -1.055221470 -1.140124092 -0.726906145
#>   0.488061433 -0.185417684 -0.586684275
#>  -1.258520691 -1.201624748  2.025176081
#>   2.408952918  3.855327394 -1.687018757
#>   1.403728361 -3.360111869 -1.182513461
#>   0.968488611  1.740950812 -1.598891631
#>   0.830086522 -0.945981855  0.601873796
#>   1.552672544  2.177841077 -2.362245020
#>  -0.345491394 -1.251718930  1.127413365
#>  -2.764665838 -0.633756820  2.565270365
#>  -0.301609077  1.860566508 -2.007024070
#>  -1.619915069 -2.036815823  2.178628515
#>  -0.592023575  0.971941603  0.974067365
#>   0.852399690  2.891353987 -2.871542427
#>   0.548563122 -1.457350303 -0.746670465
#>  -1.220670284  0.647567863 -0.082299486
#>   1.427445129  1.391979728  0.355938677
#>  -0.604574948 -0.791202744 -4.888084836
#>   2.042256307  0.369408364 -1.348473212
#>   0.061211561 -0.014330910 -1.027669426
#>   0.384980990 -0.847940138  0.612067814
#>   0.198839301 -3.100034648 -4.079560287
#>  -2.493809806  2.366760721 -2.159247551
#>  -2.463798345  0.788716533  4.513161691
#>  -0.803112155 -0.357624396  1.642854306
#>  -2.181139089  1.558975891  0.624673990
#>  -0.777850303  1.259389127  3.442326779
#>   3.054000026  3.446734815  3.290561957
#>   3.889108836  0.517255843 -2.848077512
#>  -0.042409688  3.622298913 -2.435030231
#>  -1.092694443 -2.098653864  1.587844449
#>  -0.861199495  0.116576838 -0.264162817
#>  -1.958114687  2.532181646  1.616876507
#>   0.266782925  2.238978612  2.906562998
#>   2.046208991  3.108554463 -0.910905219
#>   4.418778385  1.298415437 -2.859817391
#>   2.587748798 -2.440982116 -3.447985965
#>  -2.692310780 -2.912626539  0.582252499
#>  -2.962604483 -2.467279684  1.480192281
#>  -2.170745292 -3.469806047  2.019158488
#>  -0.766427015  0.315577056 -4.356863174
#>  -1.022293014 -2.129961025  2.585605368
#>   1.605742048  0.114381026 -2.244967497
#>   0.299585031 -3.414312895  1.508926847
#>   1.405492595 -1.373759197 -7.463598874
#>  -3.737306890 -4.522022001  0.844947711
#>  -3.426863952  3.378141903  0.817825753
#>  -3.110865714 -0.217420368  5.073453765
#>   0.512615135  5.064876261 -1.313999245
#>   0.546969047  0.579705912  1.212169983
#>   0.694901998 -0.090914220  2.323849257
#>  -0.816299157  0.275652234  1.918649909
#>  -0.863984639  1.929219968  3.839936321
#>   1.212050202  3.519236701 -0.897269677
#>  -1.500008437 -2.790888300 -2.342510552
#>  -2.474866839 -2.018783410 -0.500646414