Fit mixture vector autoregressive models

Provides EM-estimation of mixture autoregressive models for multivariate time series

mixVARfit(y, model, fix = FALSE, tol = 10^-6)

Arguments

y
a data matrix.

model
an object of class "mixVAR" with initial values of parameter for EM estimation.

tol
Threshold for convergence criterion.

fix
~~ TODO: describe this argument. ~~

Details

Estimation is done under the assumption of multivariate Gaussian innovations.

Value

An object of class mixVARGaussian with EM estimates of model parameters.

References

Fong PW, Li WK, Yau CW, Wong CS (2007). “On a Mixture Vector Autoregressive Model.” The Canadian Journal of Statistics / La Revue Canadienne de Statistique, 35(1), 135--150. ISSN 03195724, doi: 10.1002/cjs.5550350112 .

Examples

AR <- list()
AR[[1]] <- array(c(0.5,-0.3,-0.6,0,0,0.5,0.4,0.5,-0.3), dim = c(3,3,1))
AR[[2]] <- array(c(-0.5,0.3,0,1,0,-0.5,-0.4,-0.2, 0.5), dim = c(3,3,1))

prob <- c(0.75, 0.25)
shift <- cbind(c(0,0,0), c(0,0,0))

Sigma1 <- cbind(c(1,0.5, -0.4), c(0.5,2, 0.8), c(-0.4,0.8,4))
Sigma2 <- cbind(c(1,0.2, 0), c(0.2,  2, -0.15), c(0,-0.15,4))
Sigma <- array(c(Sigma1, Sigma2), dim = c(3,3,2))

m <- new("mixVARGaussian", prob = prob, vcov = Sigma, arcoef = AR, shift = shift)
y <- mixVAR_sim(m, n = 500, init = matrix(rep(0,3), ncol = 3), nskip = 100, flag = FALSE)

mixVARfit(y, m, tol = 10^-6)

#> niter: 25 	vallogf: -2380.11 

#> $model
#> (To see the internal structure of the object, use function 'str'.)
#> 
#> An object of class "mixVARGaussian"
#> Number of components: 2 
#> Number of series: 3 
#>           prob       shift order        ar_1                               vcov
#> Comp_1 0.71608 -0.09353062     1  0.51561572 -0.014725061  0.3707659  1.0615840
#>                -0.05341461       -0.25757624 -0.005136189  0.5002469  0.5007775
#>                 0.19589391       -0.62323339  0.517398132 -0.2976830 -0.3258831
#> Comp_2 0.28392  0.09728988     1 -0.51037371  0.959050944 -0.4377638  1.1701149
#>                -0.06480032        0.34178938  0.011680350 -0.1672338  0.2820055
#>                -0.10351943        0.07106622 -0.518795562  0.5807097  0.1609566
#>                             
#> Comp_1  0.5007775 -0.3258831
#>         1.6306777  0.7720903
#>         0.7720903  3.8258741
#> Comp_2  0.2820055  0.1609566
#>         2.1170483 -0.3208873
#>        -0.3208873  3.3933158
#> 
#> Distributions of the error components:
#> 	Multivariate standard Gaussian
#> 
#> 
#> $vallogf
#> [1] -2380.198
#>