Exact predictive parameters for multi-step MixAR prediction

Exact predictive parameters for multi-step MixAR prediction.

Usage

predict_coef(model, maxh)

Arguments

model

a MixAR model.

maxh

maximal horizon.

Details

predict_coef() implements the method of boshnakov2009mar;textualmixAR for the h-step prediction of MixAR processes. The h-step predictive distribution has a MixAR distribution with \(g^h\) components and this function computes its parameters.

predict_coef() implements the results by boshnakov2009mar;textualmixAR to compute the parameters of the predictive distributions. predict_coef() is mostly a helper function, use multiStep_dist for prediction/forecasting (the exact method for multiStep_dist uses predict_coef() to do the main work).

predict_coef() returns a list of lists containing the quantities needed for each horizon \(h\), see section Value.

Alternatiely, the parameters can be obtained as MixAR models by calling the function generated by the exact method of multiStep_dist with argument what = "MixAR".

Value

a list with components:

arcoefs

a list, arcoefs[[h]] gives the ar coefficients for the h-step predictive distribution.

sigmas

a list, sigmas[[h]] sigmas[[h]] is a matrix, in which the \(k\)th column contains the theta coefficients needed to compute \(sigma_k\) in the formula for sigma in Equation (16) @see @boshnakov2009marmixAR. In the paper the index is a tuple \((k_1,…,k_h)\) for clarity. In the code each tuple \((k_1,…,k_h)\) is mapped to a linear index in \(1,\ldots,g^h\) (there are \(g^h\) tuples for horizon \(h\), since the mixture has \(g^h\) components).

probs

a list, probs[[h]] gives the mixture weights for the h-step predictive distribution.

sStable

a list, sigmas[[h]] gives the scale parameters for the h-step predictive distribution.

References

Author

Georgi N. Boshnakov

See also