Create objects from class "raggedCoef"

Create objects from class "raggedCoef".

raggedCoef(p, value = NA_real_)

Arguments

p
orders, vector of integers.

value
typically, a list, but see Details.

Details

This is an alternative way to create raggedCoef objects. It is more consistent than the "initialize" method that underlies new("raggedCoef", ...).

If value is a "raggedCoef" object it is returned. If value is a list, it is converted to "raggedCoef" using new(). if argument p is not missing, a consistency check is made to ensure that the order of the object is as specified by p.

Otherwise, if value is of length one, it is replicated to form a ragged list with i-th element a vector of length p[i]. Although not checked, the intention here is that value is from some atomic class. The default for value is NA_real_ to give a convenient way to create a ragged list.

Finally, if none of the above applies, value is effectively assumed to be a vector of length sum(p), although other cases are admissible (but I don't remmeber if this is intended). In this case, value is reshaped into a ragged list to match p. This is convenient when, for example, the elements of a ragged array are obtained from an optimisation routine which expects plain vector. (TODO: the description of this case is from cursory glance at the code and guesswork - check.)