MixAR models for examples and testing

MixAR models for examples and testing.

exampleModels

Details

Coefficients of models from the examples in Wong and Li (2000) . Variations on these with different noise distributions are used throughout the examples in mixAR. The models are from classes inheriting from class "MixAR".

exampleModels is a list with the following components:

WL_ibm
WL_A
WL_B
WL_I
WL_II
WL_ibm_gen
WL_ibm_t3v
WL_ibm_tf
WL_At
WL_Bt_1
WL_Bt_2
WL_Bt_3
WL_Ct_1
WL_Ct_2
WL_Ct_3

Each component is a MixAR model, i.e. an object inheriting from class "MixAR".

Source

Wong CS, Li WK (2000). “On a mixture autoregressive model.” J. R. Stat. Soc., Ser. B, Stat. Methodol. , 62(1), 95--115.

Examples

## use these instead of moWL, moWL_A, moWL_B, etc.
exampleModels$WL_ibm

#> (To see the internal structure of the object, use function 'str'.)
#> 
#> An object of class "MixARGaussian"
#> Number of components: 3 
#>        prob   shift scale   order ar_1   ar_2   
#> Comp_1 0.5439   0    4.8227   2   0.6792  0.3208
#> Comp_2 0.4176   0    6.0082   2   1.6711 -0.6711
#> Comp_3 0.0385   0   18.1716   1   1.0000        
#> 
#> Distributions of the error components:
#> 	standard Gaussian
#> 

exampleModels$WL_A

#> (To see the internal structure of the object, use function 'str'.)
#> 
#> An object of class "MixARGaussian"
#> Number of components: 2 
#>        prob shift scale order ar_1
#> Comp_1 0.5    0     5     1   0.5 
#> Comp_2 0.5    0     1     1   1.1 
#> 
#> Distributions of the error components:
#> 	standard Gaussian
#> 

exampleModels$WL_B

#> (To see the internal structure of the object, use function 'str'.)
#> 
#> An object of class "MixARGaussian"
#> Number of components: 2 
#>        prob shift scale order ar_1
#> Comp_1 0.75   0     5     1   0.5 
#> Comp_2 0.25   0     1     1   1.4 
#> 
#> Distributions of the error components:
#> 	standard Gaussian
#> 

# what is the difference between A and B?
show_diff(exampleModels$WL_A, exampleModels$WL_B)

#>        prob  shift scale order ar_1
#> Comp_1 -0.25   0     0     1    0.0
#> Comp_2  0.25   0     0     1   -0.3

exampleModels$WL_I

#> (To see the internal structure of the object, use function 'str'.)
#> 
#> An object of class "MixARGaussian"
#> Number of components: 3 
#>        prob shift scale order ar_1 ar_2  ar_3
#> Comp_1 0.4    0     1     2    0.9 -0.60     
#> Comp_2 0.3    0     1     1   -0.5           
#> Comp_3 0.3   -5     5     3    1.5 -0.74 0.12
#> 
#> Distributions of the error components:
#> 	standard Gaussian
#> 

exampleModels$WL_II

#> (To see the internal structure of the object, use function 'str'.)
#> 
#> An object of class "MixARGaussian"
#> Number of components: 3 
#>        prob shift scale order ar_1 ar_2
#> Comp_1 0.4    5     1     2    0.9 -0.6
#> Comp_2 0.3    0     1     2   -0.7  0.0
#> Comp_3 0.3   -5     5     2    0.0  0.8
#> 
#> Distributions of the error components:
#> 	standard Gaussian
#> 

#show_diff(exampleModels$WL_I, exampleModels$WL_II)

exampleModels$WL_ibm_gen

#> (To see the internal structure of the object, use function 'str'.)
#> 
#> An object of class "MixARgen"
#> Number of components: 3 
#>        prob   shift scale   order ar_1   ar_2   
#> Comp_1 0.5439   0    4.8227   2   0.6792  0.3208
#> Comp_2 0.4176   0    6.0082   2   1.6711 -0.6711
#> Comp_3 0.0385   0   18.1716   1   1.0000        
#> 
#> Distributions of the error components:
#> [[1]]
#> [[1]]$pdf
#> [1] "dnorm"
#> 
#> [[1]]$cdf
#> [1] "pnorm"
#> 
#> [[1]]$rand
#> [1] "rnorm"
#> 
#> [[1]]$logpdf
#> function (x) 
#> log(dnorm(x))
#> <bytecode: 0x7f8ea1a07070>
#> <environment: 0x7f8ea1a0dcb0>
#> 
#> [[1]]$Fscore
#> function (x) 
#> -x
#> <bytecode: 0x7f8ea1a07700>
#> <environment: 0x7f8ea1a0dcb0>
#> 
#> [[1]]$xFscore
#> function (x) 
#> -x^2
#> <bytecode: 0x7f8ea1a07cb0>
#> <environment: 0x7f8ea1a0dcb0>
#> 
#> [[1]]$Parscore
#> function () 
#> 0
#> <bytecode: 0x7f8ea1a082d0>
#> <environment: 0x7f8ea1a0dcb0>
#> 
#> [[1]]$get_param
#> function () 
#> numeric()
#> <bytecode: 0x7f8ea1a08848>
#> <environment: 0x7f8ea1a0dcb0>
#> 
#> [[1]]$set_param
#> function () 
#> stop("Standard normal has no parameters to set.")
#> <bytecode: 0x7f8ea1a04e78>
#> <environment: 0x7f8ea1a0dcb0>
#> 
#> [[1]]$any_param
#> function () 
#> param_flag
#> <bytecode: 0x7f8ea1a05c78>
#> <environment: 0x7f8ea1a0dcb0>
#> 
#> 
#> 

exampleModels$WL_ibm_t3v

#> (To see the internal structure of the object, use function 'str'.)
#> 
#> An object of class "MixARgen"
#> Number of components: 3 
#>        prob   shift scale   order ar_1   ar_2   
#> Comp_1 0.5439   0    4.8227   2   0.6792  0.3208
#> Comp_2 0.4176   0    6.0082   2   1.6711 -0.6711
#> Comp_3 0.0385   0   18.1716   1   1.0000        
#> 
#> Distributions of the error components:
#> [[1]]
#> [[1]]$pdf
#> function (x) 
#> dstd(x, nu = nu)
#> <bytecode: 0x7f8ea1a027d8>
#> <environment: 0x7f8ea1a065a8>
#> 
#> [[1]]$cdf
#> function (x) 
#> pstd(x, nu = nu)
#> <bytecode: 0x7f8ea19fefc8>
#> <environment: 0x7f8ea1a065a8>
#> 
#> [[1]]$rand
#> function (n) 
#> rstd(n, nu = nu)
#> <bytecode: 0x7f8ea19ff690>
#> <environment: 0x7f8ea1a065a8>
#> 
#> [[1]]$logpdf
#> function (x) 
#> log(dstd(x, nu = nu))
#> <bytecode: 0x7f8ea19ffd58>
#> <environment: 0x7f8ea1a065a8>
#> 
#> [[1]]$Fscore
#> function (x) 
#> -x * (1 + nu)/(x^2 + nu - 2)
#> <bytecode: 0x7f8ea1a00490>
#> <environment: 0x7f8ea1a065a8>
#> 
#> [[1]]$xFscore
#> function (x) 
#> -x^2 * (1 + nu)/(x^2 + nu - 2)
#> <bytecode: 0x7f8ea19fd0e0>
#> <environment: 0x7f8ea1a065a8>
#> 
#> [[1]]$Parscore
#> function (x) 
#> param_score_stdt(x, nu)
#> <bytecode: 0x7f8ea19fe490>
#> <environment: 0x7f8ea1a065a8>
#> 
#> [[1]]$get_param
#> function () 
#> nu
#> <bytecode: 0x7f8ea19feb90>
#> <environment: 0x7f8ea1a065a8>
#> 
#> [[1]]$set_param
#> function (x) 
#> nu <<- x
#> <bytecode: 0x7f8ea19fb1c0>
#> <environment: 0x7f8ea1a065a8>
#> 
#> [[1]]$any_param
#> function () 
#> param_flag
#> <bytecode: 0x7f8ea19fb850>
#> <environment: 0x7f8ea1a065a8>
#> 
#> 
#> 

exampleModels$WL_ibm_tf

#> (To see the internal structure of the object, use function 'str'.)
#> 
#> An object of class "MixARgen"
#> Number of components: 3 
#>        prob   shift scale   order ar_1   ar_2   
#> Comp_1 0.5439   0    4.8227   2   0.6792  0.3208
#> Comp_2 0.4176   0    6.0082   2   1.6711 -0.6711
#> Comp_3 0.0385   0   18.1716   1   1.0000        
#> 
#> Distributions of the error components:
#> 	Component 1: Student t with 20 df
#> 	Component 2: Student t with 30 df
#> 	Component 3: Student t with 40 df
#> 

#show_diff(exampleModels$WL_ibm_gen, exampleModels$WL_ibm_t3v)

exampleModels$WL_At

#> (To see the internal structure of the object, use function 'str'.)
#> 
#> An object of class "MixARgen"
#> Number of components: 2 
#>        prob shift scale order ar_1
#> Comp_1 0.5    0     1     1   -0.5
#> Comp_2 0.5    0     2     1    1.1
#> 
#> Distributions of the error components:
#> 	Component 1: Student t with 4 df
#> 	Component 2: Student t with 8 df
#> 

exampleModels$WL_Bt_1

#> (To see the internal structure of the object, use function 'str'.)
#> 
#> An object of class "MixARgen"
#> Number of components: 3 
#>        prob shift scale order ar_1 ar_2  ar_3
#> Comp_1 0.3    5   2.0     2    0.5  0.24     
#> Comp_2 0.3   -5   1.0     1   -0.9           
#> Comp_3 0.4    0   0.5     3    1.5 -0.74 0.12
#> 
#> Distributions of the error components:
#> 	Component 1: Student t with 4 df
#> 	Component 2: Student t with 6 df
#> 	Component 3: Student t with 10 df
#> 

exampleModels$WL_Bt_2

#> (To see the internal structure of the object, use function 'str'.)
#> 
#> An object of class "MixARgen"
#> Number of components: 3 
#>        prob shift scale order ar_1 ar_2  ar_3
#> Comp_1 0.3    5   2.0     2    0.5  0.24     
#> Comp_2 0.3   -5   1.0     1   -0.9           
#> Comp_3 0.4    0   0.5     3    1.5 -0.74 0.12
#> 
#> Distributions of the error components:
#> 	Component 1: Student t with 4 df
#> 	Component 2: Student t with 4 df
#> 	Component 3: Student t with 10 df
#> 

exampleModels$WL_Bt_3

#> (To see the internal structure of the object, use function 'str'.)
#> 
#> An object of class "MixARgen"
#> Number of components: 3 
#>        prob shift scale order ar_1 ar_2  ar_3
#> Comp_1 0.3    5   2.0     2    0.5  0.24     
#> Comp_2 0.3   -5   1.0     1   -0.9           
#> Comp_3 0.4    0   0.5     3    1.5 -0.74 0.12
#> 
#> Distributions of the error components:
#> 	Component 1: Student t with 4 df
#> 	Component 2: Student t with 4 df
#> 	Component 3: Student t with 10 df
#> 

## what is different between Bt_2 and Bt_1? (df of component 2)
show_diff(exampleModels$WL_Bt_2, exampleModels$WL_Bt_1)

#>        prob shift scale order ar_1 ar_2 ar_3
#> Comp_1   0    0     0     2     0    0      
#> Comp_2   0    0     0     1     0           
#> Comp_3   0    0     0     3     0    0    0 
#> 
#> Distributions of the error components:
#> 
#> 	Model 1:
#> 	Component 1: Student t with 4 df
#> 	Component 2: Student t with 4 df
#> 	Component 3: Student t with 10 df
#> 
#> 	Model 2:
#> 	Component 1: Student t with 4 df
#> 	Component 2: Student t with 6 df
#> 	Component 3: Student t with 10 df

exampleModels$WL_Ct_1

#> (To see the internal structure of the object, use function 'str'.)
#> 
#> An object of class "MixARgen"
#> Number of components: 3 
#>        prob shift scale order ar_1 ar_2  ar_3
#> Comp_1 0.3    5   2.0     2    0.5  0.24     
#> Comp_2 0.3   -5   1.0     1   -0.9           
#> Comp_3 0.4    0   0.5     3    1.5 -0.74 0.12
#> 
#> Distributions of the error components:
#> 	Component 1: Student t with 4 df
#> 	Component 2: Student t with 7 df
#> 	Component 3: Standard normal distribution
#> 

exampleModels$WL_Ct_2

#> (To see the internal structure of the object, use function 'str'.)
#> 
#> An object of class "MixARgen"
#> Number of components: 3 
#>        prob shift scale order ar_1 ar_2  ar_3
#> Comp_1 0.3    5   2.0     2    0.5  0.24     
#> Comp_2 0.3   -5   1.0     1   -0.9           
#> Comp_3 0.4    0   0.5     3    1.5 -0.74 0.12
#> 
#> Distributions of the error components:
#> 	Component 1: Student t with 4 df
#> 	Component 2: Student t with 7 df
#> 	Component 3: Standard normal distribution
#> 

exampleModels$WL_Ct_3

#> (To see the internal structure of the object, use function 'str'.)
#> 
#> An object of class "MixARGaussian"
#> Number of components: 3 
#>        prob shift scale order ar_1 ar_2  ar_3
#> Comp_1 0.3    5   2.0     2    0.5  0.24     
#> Comp_2 0.3   -5   1.0     1   -0.9           
#> Comp_3 0.4    0   0.5     3    1.5 -0.74 0.12
#> 
#> Distributions of the error components:
#> 	standard Gaussian
#> 

## The models were created with something like:
moWLprob <- c(0.5439,0.4176,0.0385)
moWLsigma <- c(4.8227,6.0082,18.1716)
moWLar <- list(c(0.6792,0.3208), c(1.6711,-0.6711), 1)

moWL <- new("MixARGaussian", prob = moWLprob, scale = moWLsigma,
            arcoef = moWLar)
moWLgen <- new("MixARgen", prob = moWLprob, scale = moWLsigma,
               arcoef = moWLar, dist = list(dist_norm))
## clean up a bit
rm(moWLprob, moWLsigma, moWLar, moWL, moWLgen)