Fit mixture vector autoregressive models

Estimate a mixVAR model for a multivariate time series. This is a generic function. The methods defined in package mixAR are described here.

fit_mixVAR(x, model, fix, ...)

Arguments

x
a multivariate time series (currently a numeric matrix).

model
model, object inheriting from mixVAR class.

fix
~~ TODO: describe this argument. ~~

...
additional arguments for the methods (not currently used).

Value

a mixVAR model.

Details

model specifies the model to fit. If model inherits from "mixVAR", it is used as a template. Estimation is done via EM-Algorithm, using the function mixVARfit.

Currently the default method for fit_mixAR just throws error, since there seems no suitable default task to do.

TODO: Define more methods to handle different cases (fix = "shift", etc..)

Methods

signature(x = "ANY", model = "mixVAR")

This is currently the only available method for fit_mixVAR

.
signature(x = "ANY", model = "ANY")

See also

Examples

AR <- list()
AR[[1]] <- array(c(0.5,-0.3,-0.6,0,0,0.5,0.4,0.5,-0.3),dim=c(3,3,1))
AR[[2]] <- array(c(-0.5,0.3,0,1,0,-0.5,-0.4,-0.2, 0.5),dim=c(3,3,1))

prob <- c(0.75,0.25)
shift <- cbind(c(0,0,0),c(0,0,0))

Sigma1 <- cbind(c(1,0.5, -0.4),c(0.5,2, 0.8),c(-0.4,0.8,4))
Sigma2 <- cbind(c(1,0.2, 0),c(0.2,  2, -0.15),c(0,-0.15,4))
Sigma <- array(c(Sigma1, Sigma2), dim=c(3,3,2))

m <- new("mixVARGaussian", prob=prob, vcov=Sigma, arcoef=AR, shift=shift)
y <- mixVAR_sim(m, n=500, init=matrix(rep(0,3), ncol=3), nskip=100, flag=FALSE)

fit_mixVAR(y, m)

#> niter: 25 	vallogf: -2146.711 

#> $model
#> (To see the internal structure of the object, use function 'str'.)
#> 
#> An object of class "mixVARGaussian"
#> Number of components: 2 
#> Number of series: 3 
#>             prob        shift order       ar_1                       
#> Comp_1 0.7879972 -0.013037900     1  0.5544802 -0.04024179  0.3944199
#>                   0.097016118       -0.2517435 -0.05989595  0.4629539
#>                  -0.010446198       -0.7203757  0.59554529 -0.3457813
#> Comp_2 0.2120028 -0.060551429     1 -0.5387120  1.00546448 -0.4048044
#>                   0.008486125        0.1116701  0.05059122 -0.2351755
#>                   0.151919121       -0.1976764 -0.52200687  0.6025687
#>              vcov                     
#> Comp_1  0.9103361 0.4801007 -0.3273280
#>         0.4801007 1.9504200  0.9809687
#>        -0.3273280 0.9809687  3.7752622
#> Comp_2  0.9139160 0.1732964  0.6322547
#>         0.1732964 1.8190884  0.1691766
#>         0.6322547 0.1691766  5.0361521
#> 
#> Distributions of the error components:
#> 	Multivariate standard Gaussian
#> 
#> 
#> $vallogf
#> [1] -2146.709
#>