A posteriori relabelling of a Markov chain

Takes the output from a MCMC simulation of parameters of a mixture, and detects whether labels switch has occured while sampling, using the method by Celeux (2000).

label_switch(x, m)

Arguments

x
output from an MCMC sampling of a mixture. A matrix, each column corresponds to one component of the mixture.

m
the number of observations in the sample that will be used to initialise the algorithm. m~100 is recommended.

Details

Function can be directly executed when x is one of mix_weights, scale, precision, shift or mu from bayes_mixAR output. ARcoeff cannot be input as it is, but element from the list may be used.

Value

A list of 2:

x

The input matrix, with adjusted labels

true_perm

The "true" permutation at each iteration.

References

Celeux G (2000). Bayesian Inference of Mixture: The Label Switching Problem.. Payne R., Green P. (eds) COMPSTAT. Physica, Heidelberg.

Note

There is no absolute choice on what x should be to obtain the "true" permutation at any given point. User is subject to make the most suitable choice, given output of their MCMC.

See also

Examples

model <- new("mixARGaussian",
             prob   = exampleModels$WL_At@prob,      # c(0.5, 0.5)
             scale  = exampleModels$WL_At@scale,     # c(1, 2)        
             arcoef = exampleModels$WL_At@arcoef@a ) # list(-0.5, 1.1)

y <- mixAR_sim(model, n = 300, init = rep(0, which.max(model@order)))

## just examples, use larger numbers in practice
nsim <- 30   # 200
burnin <- 10  # 100
x <- bayes_mixAR(y, model, fix_shift = FALSE, tau = c(.15, .25),
              nsim = nsim, burnin = burnin)

label_switch(x$mix_weights, m = 5)

#> $x
#>            [,1]      [,2]
#>  [1,] 0.5476954 0.4523046
#>  [2,] 0.5395222 0.4604778
#>  [3,] 0.5730514 0.4269486
#>  [4,] 0.6110079 0.3889921
#>  [5,] 0.5771588 0.4228412
#>  [6,] 0.5515905 0.4484095
#>  [7,] 0.5116267 0.4883733
#>  [8,] 0.5180564 0.4819436
#>  [9,] 0.5686049 0.4313951
#> [10,] 0.5590577 0.4409423
#> [11,] 0.5669298 0.4330702
#> [12,] 0.6118508 0.3881492
#> [13,] 0.5704141 0.4295859
#> [14,] 0.5977270 0.4022730
#> [15,] 0.6357773 0.3642227
#> [16,] 0.6169200 0.3830800
#> [17,] 0.5838272 0.4161728
#> [18,] 0.5678565 0.4321435
#> [19,] 0.6333747 0.3666253
#> [20,] 0.6100989 0.3899011
#> 
#> $true_perm
#>       [,1] [,2]
#>  [1,]    1    2
#>  [2,]    1    2
#>  [3,]    1    2
#>  [4,]    1    2
#>  [5,]    1    2
#>  [6,]    1    2
#>  [7,]    1    2
#>  [8,]    1    2
#>  [9,]    1    2
#> [10,]    1    2
#> [11,]    1    2
#> [12,]    1    2
#> [13,]    1    2
#> [14,]    1    2
#> [15,]    1    2
#>