Fit mixture autoregressive models with seasonal AR parameters

Provides estimation via EM-Algorithm for mixture autoregressive models including seasonal AR parameters.

mixSARfit(y, model, est_shift = FALSE, tol = 10^-14)

Arguments

y
a time series (currently a numeric vector).

model
an object of class "mixAR" including seasonal components.

est_shift
if missing or FALSE, fix the intercepts to zero, otherwise estimate them.

tol
threshold for stopping criterion.

Details

This function only works for "mixAR" objects in which slot arcoef is of class "raggedCoefS".

Value

A list of 2:

model

an object of class "mixAR". The estimated model.

vallogf

the value of the loglikelihood function for the returned model.

Examples

ar1 <- list(c(0.5, -0.5), c(1.1, 0, -0.5))
ar12 <- list(0, c(-0.3, 0.1))
s = 12

rag <- new("raggedCoefS", a = ar1, as = ar12, s = s)

model <- new("mixARGaussian", prob = exampleModels$WL_A@prob, # c(0.5, 0.5)
             scale = exampleModels$WL_A@scale,                # c(5, 1)
             arcoef = rag)

#set.seed(1234)
y <- mixAR_sim(model, n = 500, init = rep(0, 24))

mixSARfit(y, model)

#> Initial vallogf:  -1289.877 
#> Final niter:  21 	vallogf:  -1285.272 

#> $model
#> (To see the internal structure of the object, use function 'str'.)
#> 
#> An object of class "mixARGaussian"
#> Number of components: 2 
#>        prob      shift       scale     order ar_1      ar_2         ar_3      
#> Comp_1 0.4960326 -0.25384321 4.8804343   2   0.4653227 -0.497871220           
#> Comp_2 0.5039674 -0.03347416 0.9368791   3   1.0984148  0.003608709 -0.5104407
#>        s_order ar_12       ar_24     
#> Comp_1 1       -0.03180571           
#> Comp_2 2       -0.30409580 0.09912557
#> 
#> Distributions of the error components:
#> 	standard Gaussian
#> 
#> 
#> $vallogf
#> [1] -1285.272
#> 

## fix the intercepts to zero
mixSARfit(y, model, est_shift = FALSE)

#> Initial vallogf:  -1289.877 
#> Final niter:  21 	vallogf:  -1285.272 

#> $model
#> (To see the internal structure of the object, use function 'str'.)
#> 
#> An object of class "mixARGaussian"
#> Number of components: 2 
#>        prob      shift       scale     order ar_1      ar_2         ar_3      
#> Comp_1 0.4960326 -0.25384321 4.8804343   2   0.4653227 -0.497871220           
#> Comp_2 0.5039674 -0.03347416 0.9368791   3   1.0984148  0.003608709 -0.5104407
#>        s_order ar_12       ar_24     
#> Comp_1 1       -0.03180571           
#> Comp_2 2       -0.30409580 0.09912557
#> 
#> Distributions of the error components:
#> 	standard Gaussian
#> 
#> 
#> $vallogf
#> [1] -1285.272
#>