Simulate from multivariate mixAR models

Simulate data from multivariate mixAR models under the assumptions of multivariate Gaussian innovarion

mixVAR_sim(model, n, init, nskip = 100, flag = FALSE)

Arguments

model
model from which to simulate, an object inheriting from class mixVAR.

n
size of simulated multivariate series.

init
initial values, a numeric matrix. If missing, a matrix of 0 values is generated.

nskip
number of burn-in values.

flag
if TRUE returns also the regimes.

Details

mixVAR_sim simulates a series of length nskip + n and returns the last n values. init provides initial values for the algorithm. Each row is considered as a time point. The number of rows must be at least equal to the maximal AR order.

Value

a numeric matrix with n rows.

See also

Examples

AR <- list()
AR[[1]] <- array(c(0.5,-0.3,-0.6,0,0,0.5,0.4,0.5,-0.3), dim = c(3,3,1))
AR[[2]] <- array(c(-0.5,0.3,0,1,0,-0.5,-0.4,-0.2, 0.5), dim = c(3,3,1))

prob <- c(0.75, 0.25)
shift <- cbind(c(0,0,0), c(0,0,0))

Sigma1 <- cbind(c(1, 0.5, -0.4), c(0.5, 2, 0.8), c(-0.4, 0.8, 4))
Sigma2 <- cbind(c(1,0.2, 0), c(0.2, 2, -0.15), c(0, -0.15, 4))
Sigma <- array(c(Sigma1, Sigma2), dim = c(3,3,2))

m <- new("mixVARGaussian", prob=prob, vcov=Sigma, arcoef=AR, shift=shift)
mixVAR_sim(m, n=500, init=matrix(rep(0,3), ncol=3), nskip=100, flag=FALSE)

#>          [,1]          [,2]         [,3]
#>   0.489631461 -0.0806328015  0.152611628
#>  -0.839739034 -2.1558373576 -0.387692069
#>  -1.291398079 -0.6334534627 -3.570422951
#>   0.538481685  0.5874366203  0.608544604
#>   0.510248815 -1.2868953605 -1.308018410
#>   0.779502508  0.3538536794  0.958872326
#>   0.606525550  0.6087088901  2.505335722
#>   0.492767625  1.1725925637 -1.256445116
#>   0.165211122  1.2156516479  1.963431841
#>  -0.310041606 -0.8476689162 -1.513684220
#>  -1.085400910 -1.1729204222  1.671586580
#>  -0.295477774  2.2024965407  1.243024564
#>   0.256626525  2.5745414870  1.382542840
#>   0.225844592  0.3184338108 -1.121633138
#>  -0.118587728 -1.3845011653  2.593390165
#>  -1.218290014 -0.7582981696 -2.022975865
#>   0.073609491  1.8841985425 -0.813296178
#>   0.191809984  0.6623872282  1.565569535
#>   0.935329467  0.4140291350  1.829501314
#>  -2.362609203 -1.7001264257 -3.696243884
#>  -2.420028142  0.5235246720  0.137454308
#>  -0.787742656  1.5323243343  2.205352343
#>   0.171692784  0.2659512208 -0.792438167
#>   0.209351922  1.1424221590 -0.980575230
#>  -0.505552534 -0.2476915868  2.349240357
#>   0.845100895  1.7748988374 -3.291103102
#>  -0.011051293 -1.8742008421 -0.887581226
#>  -2.787219008 -1.3133291358  0.930732193
#>  -0.806789284  0.8878000271  0.694958631
#>  -0.157859423 -0.3937156666  0.615417946
#>   0.077261047 -0.5618246116 -0.092489787
#>   0.310734039  0.9021912417  1.190053167
#>   0.704621643  0.3188021838 -0.059646817
#>   0.364435134 -1.7166992647  0.662323982
#>   0.566124374  3.1583576906 -1.788988574
#>  -2.795772942  0.0029469517  1.383863626
#>  -0.393440557  2.2509856416  3.396550047
#>   0.091442837 -2.5971317989  2.753432486
#>  -0.248890085 -0.7833917776 -0.873770867
#>  -0.184751660  3.2942805552  0.541519127
#>  -0.622064188 -0.1960728083  4.826653894
#>   0.507807159  3.3064093393  0.281456756
#>   0.287451114  1.7777298369  0.257333607
#>  -0.044278323  1.1495914857  5.334676681
#>  -0.206127473  1.4463196736  0.560315389
#>  -1.096314595  0.7534963178 -2.049129108
#>   0.834419251  0.8337078672 -1.541357355
#>   0.803023155  1.3696977896  2.409887956
#>   1.837601983  1.8462465269 -0.155253021
#>   2.383889583  0.0389201970  1.210916375
#>  -2.573433244  0.5381579564 -3.411162940
#>  -2.840970159 -2.9776836030  2.180749211
#>  -0.454301674  3.0613295531 -1.429108545
#>  -1.988962237 -3.2853360865  3.070480946
#>   0.996580048  2.4422822912  1.331934720
#>   0.919874640 -1.4259058964 -0.053381713
#>  -1.306675306 -1.6918351666 -2.517644095
#>  -1.139770143 -0.7814546354 -3.832001660
#>  -0.352397074  0.6778751306  0.236137115
#>   0.478996809  1.5363109168 -1.501440240
#>  -0.196751781  0.8995887212  1.132521350
#>   0.161220910 -0.5873142589  0.132136889
#>  -0.749758662 -2.8433679190 -2.698491529
#>  -1.812004657 -0.6883705374  2.028555983
#>  -0.259243525  1.6678252615  2.284887221
#>   2.900640425  1.9545597949  2.236967231
#>   0.085290565  0.4404278366 -3.695695188
#>   0.864238608 -0.0524214999 -4.264616321
#>  -2.632613152 -2.9891481436  2.695536701
#>  -1.062951753  2.0366201715 -0.952467909
#>   3.742428332 -0.0008534025 -1.831512107
#>   0.606107087 -0.6648289080  0.870937735
#>  -1.025027937  4.6261530490  0.617930028
#>   2.756438663  0.7161117892 -0.853253480
#>  -0.762946650 -2.4101018706  1.889309227
#>  -1.053505945  0.0822684830  2.751126778
#>   0.013507111  0.4596092088 -1.657492864
#>  -1.520069754 -1.9303688079 -0.512024601
#>  -0.528852568  1.7801520481  3.381964297
#>  -1.099412793  1.8843524177  0.067601695
#>  -1.152546613 -1.9597411236  1.148434327
#>  -0.606744228  0.9515198643 -2.624709565
#>   1.819673214 -1.4784990901  0.337227080
#>  -0.542855091 -0.5750274435  2.400116287
#>  -1.599644479  0.2273817294 -0.418859680
#>   0.729874885  1.0899011349 -0.341269145
#>   1.059588990 -0.2619218133  0.318892487
#>  -1.685448550  0.7708045612  3.434622394
#>   1.762388102  3.0164996759  0.905831319
#>   1.795732510  1.3510677146  1.101011162
#>   2.032204590  1.0841422310 -0.216841518
#>   1.477181670 -0.2200129293  1.290629425
#>   0.883236281  0.3773519406 -1.204254754
#>  -0.613841691  0.2405187048 -1.813250655
#>   0.127422909 -2.1327808099 -1.380060574
#>   0.606960171 -1.1019625174 -4.533546393
#>  -0.532905467 -5.6905200464 -1.780009088
#>  -2.583607621 -1.8388080176 -1.721019891
#>  -2.408732494 -0.9979397688  0.730570061
#>  -0.528165503 -0.5422779173 -3.632907073
#>  -3.145198000 -4.3275301863  2.061598420
#>  -1.812139876 -2.7778595109  1.476817294
#>  -0.979027650  2.1699236085 -0.349667098
#>  -0.020205230  1.3526553332  2.383876659
#>   1.114682214 -0.3098924695 -0.508703427
#>  -0.357176506 -2.8138660598 -3.213272010
#>  -0.800795236 -3.1691266799 -1.498960319
#>  -0.081844804  0.3324107057  0.565248467
#>   0.811600468 -0.3694137511 -4.394620383
#>  -2.196236347 -4.5677865189 -0.350263378
#>   0.878497912  0.9684822774 -2.569116371
#>   0.391118423 -0.6883783203  0.553117016
#>  -1.344204067 -0.2143513521  0.484067252
#>   0.243714143  0.8071064404 -0.751616023
#>   1.760947749 -0.7503600648  1.870240555
#>  -3.292030920  1.0729056307  3.791168550
#>   0.910571540  3.1841704189  2.525373661
#>  -0.374733613  0.2221406018 -0.570079243
#>   1.245906426  0.1479200773  2.188801952
#>  -1.459776073  0.9618633355  0.776079760
#>  -0.168464053 -0.5614736307  1.914650713
#>  -1.599607343  0.1297991827 -0.530002134
#>  -0.040595880  4.8308825772  4.893157294
#>   2.684993911  1.9164690100  1.307421522
#>   1.292325684 -2.9733449062 -1.990669792
#>  -0.391971159 -1.7118948266 -2.690383913
#>  -0.382969599  1.8449590598  2.849660467
#>   1.559149596 -0.5555929357  0.652160955
#>   2.540327115  1.7373722977 -2.440481936
#>   2.290157307 -0.2878630895 -4.420438013
#>   1.928607813  1.7543688241 -1.998417582
#>   2.547694740  0.8193774700 -3.559461762
#>   0.964643630 -0.9025351686 -3.058756546
#>  -1.253650450 -3.7891484668 -0.082378087
#>  -1.062449045  1.6249142471 -1.281618188
#>  -2.028157964 -0.2589261103  1.958703115
#>  -0.080758752  1.7945797090 -0.232185262
#>   1.658201230  0.8378672985  2.210385687
#>  -0.320118566  0.2211352190  0.964987716
#>  -1.083444160  1.2761460698 -0.439135869
#>   0.162805429  2.5161903637  3.454943466
#>   2.318026020 -1.3893672083 -3.134988550
#>   0.339064423 -3.5165556339 -2.554510397
#>  -1.275114237  0.8900273764  0.568199010
#>  -0.722334373 -2.0527274027 -3.041022678
#>  -1.499409540 -1.1961711046 -2.167467375
#>  -3.137295943  0.8260083960  2.554481516
#>   0.041578802  3.7270991285  2.780177657
#>   0.606767226  2.5313311245 -1.046167210
#>  -0.541443725 -0.2876028048  1.929090903
#>   1.395458564  2.2644371281 -1.872497257
#>   0.936405722 -0.9868904046 -0.869958485
#>  -1.532510069  0.9470528422  2.405118643
#>  -0.135373220  2.6156713184 -0.268984533
#>   0.557422671 -0.1168070640 -0.300186755
#>   1.371120349  0.6768498772 -1.988465643
#>   0.626014349  1.8781824316 -0.751705613
#>   0.349635381  1.0458241823 -1.038255177
#>   1.972803955 -1.1203750211 -0.944071334
#>   0.316101405  1.2096352149  5.214199830
#>   3.064905209  3.1318168876 -3.786990358
#>  -0.351466998 -6.0841565977 -0.258550618
#>   1.087664520  0.0274026478 -3.144651349
#>  -0.720684141 -2.0344976206  0.465475791
#>  -1.683440722 -0.6454059458 -1.284861400
#>  -0.642418866  0.0896229151 -0.464866265
#>   0.790769650  1.5209898003 -1.787516862
#>  -0.221180106 -2.2684256988 -0.596299726
#>  -1.430489626  1.2589833082  1.559655047
#>  -0.504838488  2.8936798177  3.426347934
#>   0.420804514  1.7767296426 -1.340088630
#>   3.587765399  0.8502443902  0.565249163
#>   2.250892800 -0.6654687157 -0.499901346
#>  -0.366761535 -3.3041896908 -2.967483124
#>  -1.706051294 -0.0579061022 -0.743130489
#>  -0.533595821 -0.0933664315  2.876671028
#>  -1.769819599 -1.1251908872 -0.156862999
#>  -0.698242856 -2.1939111258  1.428527644
#>   1.156617787  2.0129258837  2.167322427
#>  -0.385221904 -0.5429145715  0.768089931
#>   0.526617755  1.5302640568  3.165614161
#>   2.866608492 -0.2956263738 -1.951198264
#>  -2.462870348 -0.3343688726 -0.287779167
#>  -1.236258143 -0.3199922960  1.744097874
#>  -0.579536677  1.4454396943  0.457919833
#>  -0.528646031 -1.0968169508 -0.182752876
#>   0.070840675 -0.5789408231 -1.020097988
#>  -1.496430315 -1.3804930572 -0.016831400
#>  -1.385276007 -0.9977432360 -1.503070605
#>   0.002215670  2.7644443017  0.376603075
#>   1.025168919  2.3711453672  0.829359114
#>   0.768353904  0.6682160318 -1.340250560
#>  -0.295249106 -1.9414148836 -1.750498820
#>  -0.813677573 -0.3388465293  3.923339843
#>  -1.291077816  2.2300908512 -1.879314065
#>  -2.013694224 -1.9687928239  3.370666032
#>   1.898149943  3.0464246973 -2.680319801
#>   1.361882493 -1.7169055136  1.249757993
#>   1.071105337 -0.0710956922 -1.936483405
#>  -1.063119198  0.2167218432  1.629507553
#>   0.604608545  2.1448105765 -1.936777536
#>   0.540376277 -0.4825985936  2.170086205
#>   0.604930092 -0.3924656573 -2.421275442
#>   0.338986044  0.7324533516  3.353881563
#>   1.047596482  0.8903851935  1.890636782
#>  -0.341105343  0.2049998082  0.644877275
#>  -1.169696953 -2.5508391223  1.265314411
#>  -0.916899691 -1.4476082908  0.568469051
#>  -0.682261280  2.2033839033 -1.227668962
#>   2.133509105 -1.1374094994 -2.856686058
#>  -1.606381120  2.0041939019 -0.283747762
#>  -0.719097736 -0.2346780848  3.253186413
#>  -1.751648046  2.7374101293  0.488715612
#>  -0.769118184  1.0239433610  0.035135292
#>  -0.709641786 -0.1377634632 -2.002481715
#>  -1.159422119 -0.9170973597  0.403064146
#>  -1.159659513  1.4818778983 -2.329827512
#>   1.475105822 -0.8950087469  1.215843231
#>   1.420124159  0.4925114830 -2.076493598
#>   0.175660634 -0.3560344619  1.532471404
#>   1.862181994 -0.0774189318 -1.463487620
#>   0.985278539 -0.1453508372 -1.234345191
#>   0.550428994 -1.6005318386 -1.207228295
#>   0.295086297 -3.0642602249 -0.957373671
#>   0.476958299 -0.6707762776 -0.416040440
#>  -1.432654032  0.4087493180  0.138523482
#>  -0.957703590 -1.2994358741  0.386954508
#>  -1.348356796  0.0546708605  1.872079478
#>  -0.493117020  1.1904638948  0.558598763
#>  -0.012309272  1.4794443191  0.870146071
#>   0.344586148 -0.8429499965 -0.516360665
#>  -0.725431367  0.4273485643  2.023775188
#>  -0.663586016 -3.7493698006 -0.795470380
#>  -3.932147553  1.5422390240  1.709762089
#>   0.554217326  2.5601780635 -0.587643141
#>   2.949185734 -0.7303946377 -1.539186850
#>   0.779618599 -1.9582466304 -2.348595393
#>   0.630316216  3.0329112447 -2.037416339
#>  -0.214768931  0.6936553899  4.186862683
#>   1.142728974  3.7105962369 -1.683116304
#>   0.357705391 -1.4024435450 -2.101270318
#>  -1.050559207 -2.3607944560 -4.188861520
#>  -1.965907548 -0.9937356603  2.669303177
#>  -0.757231734 -2.0076008060  0.260569381
#>   0.709053770  2.7344406392  1.189289357
#>   0.769068747  0.7435475241  3.755443165
#>  -0.250935454  2.3927200921 -0.912340228
#>  -1.227020260  0.2627822303  0.065608526
#>  -0.236013751  0.7702749945  1.119966184
#>  -1.392502093 -0.1977980283 -0.820682067
#>  -1.796494518 -2.0640063802 -1.455515329
#>  -1.321768586  1.4109062402  1.638326129
#>   0.403790703  4.0376229036  4.581542639
#>   2.405259445  1.1111154237  2.801241851
#>  -1.433312590 -0.2083494454  1.330461648
#>  -0.001359195 -1.4179041794  2.160625185
#>  -2.809090016 -1.2583685945  1.305069536
#>  -1.561496850  1.5065170318  0.628993029
#>  -0.422063696  0.8960399953 -1.946755231
#>   2.213303587  1.5166965833  0.299700996
#>   3.374828148 -1.1215443212 -3.105595789
#>   0.180625060 -4.5013417924  0.269877028
#>  -0.653522759  0.3566608999 -2.835815751
#>  -3.514735202 -1.6905415967 -0.325610536
#>  -1.547586423 -2.2933309601  0.346907194
#>  -1.273078267  1.7601872563 -1.316806740
#>  -0.393307351 -0.6709216870  2.033327328
#>   1.580410383  0.9896275320  2.490330295
#>   3.379957927  2.2018240944 -5.477599079
#>   2.414014065 -0.0074307199 -2.257188991
#>   1.415588409 -2.5457862431 -2.131855974
#>  -3.949305213  1.7680655341  3.092512079
#>  -2.967070598  0.1357516173  4.401825383
#>   1.620453781  3.9729659770  0.889497202
#>   0.718063446  0.1677511662  0.034231863
#>   2.719874375 -1.1259150454 -6.994066972
#>  -1.375973366 -5.7180121613  1.139505651
#>  -1.837651175  0.3676262622 -1.426743686
#>  -1.906322534 -3.2364488453 -1.564793957
#>  -2.000947316  1.7914844325  0.282162451
#>   3.964594502 -0.2756917687 -1.749111962
#>   1.870104880 -3.0085200290  0.476257278
#>  -3.192505090  0.9895143983  2.877829856
#>   1.599985660 -2.1320925237  0.277195302
#>   0.523652107  1.3690799586 -0.520207179
#>  -0.363515393 -1.9258780884  0.891993688
#>  -3.172257935 -1.6625130634  5.521770984
#>  -0.831639094  2.0331411315 -1.736313441
#>   2.456040032 -1.5316807299 -2.211279763
#>  -2.797942360  2.9817044177 -0.153182582
#>  -0.958392693 -0.9719189732  2.702733714
#>   1.145799585  1.9039584642 -0.747415096
#>  -0.801564021 -1.9203928012 -0.308061623
#>  -1.392229132  1.3886068867  3.409624811
#>   2.046943103  2.9771610125  0.173393458
#>   2.393141538 -0.4971946495 -0.836767766
#>   0.184810546 -3.8438516556  1.707664451
#>  -1.109786806 -3.1138813744 -2.993505690
#>  -1.672872126 -1.1149330867  0.266333204
#>  -1.946098641  1.5191508935  3.430043598
#>   0.309440183 -0.8959929796  0.399291677
#>  -1.189593078  1.3725051290  1.160028737
#>   1.417087607  1.2965977364  3.599947364
#>   1.708415513  0.8704618388 -1.915814137
#>   0.449531759 -0.7272952563  0.343644725
#>  -0.557530327 -0.9769297827  0.092973305
#>   0.001194465  1.3182591583  1.166579624
#>   1.247674241  2.0519574161  1.612911715
#>   0.919325376  0.3100501966  1.017806129
#>   0.413238459 -0.0761486218 -0.845542973
#>  -1.667186882 -0.3122642428  1.425736176
#>   0.039000267  1.7489357817  0.438488698
#>   0.751600909  1.2680909065  2.314555560
#>   1.570900777  0.9062268487  0.490413528
#>   1.169290959  0.3462951643  0.808258290
#>   1.708717298  0.1230457517 -1.663041728
#>   1.535643710 -1.0584722796 -1.402754559
#>  -2.185251337  1.0265587763  0.588935141
#>  -2.290150693 -1.1177867841  1.852988883
#>  -0.706492085  2.9995094313 -1.467075117
#>  -2.516688168 -0.8584422249  3.513380650
#>   0.210780633  2.4765173107  0.721564471
#>  -0.161712793  0.2237136231 -0.577649676
#>  -1.123903388 -0.1769468586  1.348047473
#>  -1.055832698 -1.5746658919 -2.100510632
#>  -2.188395882  0.0934703184  0.898212294
#>   0.862647229  0.1756659417 -0.348297011
#>   0.366979947 -0.3310232056 -1.828441669
#>   0.103564320 -1.0851339543  1.872321508
#>  -0.855340864  0.3923745262 -0.269138098
#>  -1.019281834 -0.8078468183 -0.514243675
#>  -0.164977091  1.5498329002 -1.365615972
#>  -1.055334631 -0.9660740012  3.974305243
#>   3.018808396  3.4528846138 -1.086218923
#>   2.289332332 -1.7810699415 -0.199293759
#>   2.172058931 -0.5173770064 -3.282007515
#>  -0.443675113 -2.3979035796  2.981389594
#>   0.819753790  2.8270028886 -1.530542927
#>   1.391336302  0.4118519313  0.122422532
#>   0.903561603 -0.3657894143  0.248734679
#>   0.633588862  0.4590588190 -0.906515176
#>   1.481622919  0.9506022047 -3.326496244
#>   1.493120054  0.5838692402  0.108574488
#>  -0.861220283 -3.3502145536 -0.042144654
#>   0.351157746  1.4140450316  1.496728567
#>   1.280771365 -0.1936190907  1.515596226
#>   0.781757916  2.0022821880  1.011595728
#>   1.245558847  1.5468108418 -1.078069411
#>   0.793595562 -4.1076428559 -2.463250566
#>  -0.149265368 -2.8021971320 -7.219292184
#>  -3.203547183 -1.8437604320 -1.004340579
#>  -1.960097159 -0.1759791914  1.188577179
#>  -0.687275551  0.9242179068  0.482855784
#>  -1.094547183 -0.4885031279  0.033278867
#>   0.172096418  2.3262739211  0.752085244
#>   0.301713602 -1.5480184507  0.632920947
#>  -0.572718182  2.5210639514 -1.110425027
#>  -0.086774779 -0.0223995508 -0.594012555
#>  -0.151775280 -1.5215832485  0.785757104
#>  -2.432654786 -1.2317798745  2.011050729
#>  -0.002805183  1.2945736262 -1.335418493
#>  -0.509916346 -0.4496251262  2.905364877
#>  -0.081911186  1.8572212198  3.769457605
#>   0.243046969  2.1507293397  0.938370234
#>   0.128763526 -1.0451283682 -2.341843676
#>  -2.489909081 -1.5868506421  2.735955397
#>   0.351775786  4.0989950644 -0.934405244
#>   0.330086927 -0.1778699914  2.777103031
#>  -1.184450303  0.6584520755  1.428049299
#>   1.023860475  1.9635557224  0.564082820
#>   0.435370374 -1.9655181451 -1.385784564
#>  -0.496654669  0.5974802002  1.011601740
#>  -0.998567874  0.2480760032  3.736495155
#>  -0.363115367  2.1570663730  0.416206936
#>  -0.752621294 -1.7810792212  0.128795230
#>   2.266722425  0.7047580337 -1.365039463
#>   0.372697554 -2.7381990555  0.568408713
#>   1.014993016 -0.1245724109 -1.844738392
#>  -0.297811940 -0.8102707265 -3.148608278
#>  -1.740578035 -0.2595568270  4.246258489
#>  -0.285827558  0.8729058152 -0.710964920
#>   1.305732686  1.1263611067 -1.054087995
#>   0.167669772  1.4609576815  0.005679884
#>   1.172523065  1.0681710335 -0.193217033
#>   2.118700706 -1.0611241637 -2.452820900
#>   0.740206691 -3.5714661977 -0.375356157
#>   0.458692096  0.8459112775 -0.417639104
#>  -0.861790882 -1.2365343596  1.816821252
#>  -0.741096671 -2.0653529604 -0.886411209
#>  -1.757374733 -0.4946119494 -0.200097928
#>   1.424404473  1.6650281100  0.340215064
#>   0.314197391 -1.2682751567 -1.401727786
#>  -0.917451873 -1.6035056422 -6.497201446
#>  -3.974998517 -1.8450660890  2.422498034
#>  -0.841507707  2.0739854765  3.290727811
#>   0.743386351  2.6388206034  0.774571896
#>   0.845983285 -0.0207477974  2.917246336
#>   1.778432107  4.9708973109 -0.919014574
#>   0.385374473 -1.6323406541  2.969870257
#>   1.597714322  2.4630046360 -0.398436701
#>   0.277049720 -0.0192412920 -1.635745418
#>   0.723253433  1.7808347789 -4.030387159
#>  -1.446379160 -3.4017492888 -2.445982154
#>  -1.351885986  0.7453023875 -2.500755746
#>  -2.813705694 -1.6531577556  1.889524474
#>  -0.882187840 -0.9343662231  1.206881382
#>   2.412047399  0.4653408261 -3.420665248
#>  -1.627124783 -3.2038020230 -1.062088150
#>  -0.285932948 -0.0377156291 -0.360455094
#>  -1.853855382 -0.8652463154 -0.084979030
#>  -2.383300922 -1.2814470532 -1.031486025
#>   1.194344865  0.3169009588  1.489944586
#>   0.499448481 -0.3284548546  2.967922820
#>  -1.935624226  1.2207837732 -0.908504051
#>  -1.958650621  1.6542609137  5.645152990
#>   1.067393228  3.1025969983  1.470706603
#>   0.969982321  0.4688960249  0.990780617
#>   0.584252052 -1.3008160515 -0.573127803
#>  -0.057394464 -2.9428578518  1.105425865
#>  -0.476748630 -0.9717017993 -1.596643218
#>   1.345818903 -1.4351324568 -3.193107885
#>  -0.418805390 -0.4937505174 -0.802651758
#>   0.579536660  1.0230092725 -2.950069550
#>  -0.424073020 -2.2349827174  0.708477778
#>  -0.047141226  0.2379135526  3.084532887
#>   0.378344334  1.8875446036 -0.761143028
#>   0.025073328 -0.1569907431 -1.855475071
#>   1.201722650  2.4092958605 -0.056799738
#>   1.688109395  2.3302845763 -0.812708108
#>   1.241496351  2.0364415710 -0.865729299
#>   1.358022780 -0.0252375213 -0.215988498
#>   0.745020840 -0.6981550817  0.602384619
#>  -0.408385790 -2.2492554523  0.272911972
#>  -1.017561830  0.1109138613 -0.402294017
#>  -1.221350606 -2.1615947848 -2.211354009
#>  -0.970055919 -3.1641842481  0.730793055
#>  -3.464754517 -2.3294761515  1.540459870
#>  -1.393638368  4.0148766730  5.526868706
#>   3.388820213  6.2390818377  4.504221013
#>   4.256451369  0.8687235767 -0.184402883
#>  -2.776739170  0.1274712290 -2.239844704
#>   1.443368462  1.4957322746 -0.247306391
#>   0.720653782 -0.2651663030  1.862459645
#>   2.057996035 -0.8226899588 -5.451290648
#>  -2.397138342 -4.2953545552  0.042888514
#>   0.302354538  2.1301704140  0.557969817
#>   1.289902580 -0.0715298905 -1.567280793
#>  -0.956135934 -3.5635156337 -3.095633604
#>  -2.785913740 -0.7349534101 -2.231620413
#>   2.362205715  2.9802323598 -1.246758562
#>   0.701640482 -0.2267843371  0.656011390
#>  -0.462752565 -1.9842800080 -0.781767163
#>  -1.832923216  0.5232870598  1.147357854
#>   0.750796205  2.5307805243  3.212912891
#>   1.954429383  2.4556602975 -0.445519919
#>   0.841589128 -1.0190271053 -1.965008093
#>  -1.371942960 -2.2620788979 -3.082561201
#>  -1.505312992 -0.2318556018 -0.944560757
#>  -1.582020159 -1.8908346995  2.527537309
#>   0.223930160  3.7673263849 -0.234323749
#>   1.911365166  0.6575941915  1.238319589
#>   2.631919420  0.8683890192 -2.313876526
#>   1.152496069 -0.8252697887  0.334961280
#>   0.219419493 -0.1102951887  0.576978257
#>   1.615122494  0.6900511869 -2.056925426
#>   0.566275948 -0.8494042431  2.163056889
#>  -0.644496147 -0.3468670495 -0.295918767
#>  -0.972550533  3.2936844595 -0.486650592
#>  -0.993873654 -1.6702574609  6.237823682
#>   2.535931672  2.9443819925 -1.834792140
#>   1.558111600 -1.3096362343 -4.493170446
#>  -1.702564453 -1.2414330674  1.107203197
#>  -0.990287561  0.8945471595  2.183631796
#>  -0.302218246  1.7467153857 -1.156155801
#>   0.433119526  1.2944468034  0.738257596
#>   0.390807882 -2.6021799103  0.185174502
#>  -2.521490372  0.6534048343  3.807407473
#>  -0.443973228  0.4197548426 -0.406637694
#>  -0.640623397  2.8890800411  1.120206166
#>   3.574337589  0.5470959698 -1.328607252
#>   0.413969896  3.5168373162 -4.122439002
#>  -2.788546949 -3.3101566729  3.784730145
#>  -1.942181999  1.6094979464 -3.661772393
#>   3.569201194 -2.1980936093 -0.224995573
#>   1.152173996 -1.1420623106 -0.226902961
#>  -0.605993954 -2.9363771562 -2.670675325
#>  -1.868067387 -0.9021129766 -0.739043958
#>  -2.720378990 -0.2387572683  2.245477463
#>  -1.631251456  3.1791768672  2.697793209
#>   1.563059683  1.8401693927  2.279585266
#>  -0.745974899 -0.3752632447 -1.877298014
#>  -0.521187859 -2.3012209009  1.510388579
#>  -0.782545887 -0.3204432003  1.444094175
#>   0.335779055  1.4875865643  0.327361982
#>   1.256561017  2.2538208824  0.833144292
#>   1.493877483 -2.7435898825  0.593849097
#>  -3.291497240  0.6141825254  1.915411968
#>  -1.633613023  1.2574691822  3.790668263
#>   0.853309758  3.8676527883  0.030042202
#>  -0.185277331  1.4861544033  1.736677714