Simulate from multivariate MixAR models

Simulate data from multivariate MixAR models under the assumptions of multivariate Gaussian innovarion

mixVAR_sim(model, n, init, nskip = 100, flag = FALSE)

Arguments

model
model from which to simulate, an object inheriting from class MixVAR.

n
size of simulated multivariate series.

init
initial values, a numeric matrix. If missing, a matrix of 0 values is generated.

nskip
number of burn-in values.

flag
if TRUE returns also the regimes.

Details

mixVAR_sim simulates a series of length nskip + n and returns the last n values. init provides initial values for the algorithm. Each row is considered as a time point. The number of rows must be at least equal to the maximal AR order.

Value

a numeric matrix with n rows.

See also

Examples

AR <- list()
AR[[1]] <- array(c(0.5,-0.3,-0.6,0,0,0.5,0.4,0.5,-0.3), dim = c(3,3,1))
AR[[2]] <- array(c(-0.5,0.3,0,1,0,-0.5,-0.4,-0.2, 0.5), dim = c(3,3,1))

prob <- c(0.75, 0.25)
shift <- cbind(c(0,0,0), c(0,0,0))

Sigma1 <- cbind(c(1, 0.5, -0.4), c(0.5, 2, 0.8), c(-0.4, 0.8, 4))
Sigma2 <- cbind(c(1,0.2, 0), c(0.2, 2, -0.15), c(0, -0.15, 4))
Sigma <- array(c(Sigma1, Sigma2), dim = c(3,3,2))

m <- new("MixVARGaussian", prob=prob, vcov=Sigma, arcoef=AR, shift=shift)
mixVAR_sim(m, n=500, init=matrix(rep(0,3), ncol=3), nskip=100, flag=FALSE)

#>          [,1]        [,2]         [,3]
#>  -1.545769916 -1.45850070 -0.850608929
#>   0.756059080  0.36496571  3.704456424
#>   1.573140885  2.27979019  0.726025321
#>   2.488469951  0.54035499  1.101636831
#>   2.120610354  0.30119473 -3.280034052
#>  -0.341775426 -1.91698076  0.552733076
#>  -0.852357125  1.22620065  0.206312293
#>  -0.424233525 -0.71682088  2.322568351
#>   0.582362229  1.03897812 -1.981168197
#>   0.032525457 -2.32050396  0.343610476
#>  -0.867836619 -1.82816053 -1.929793536
#>  -3.228539987 -2.09810385 -0.010749438
#>  -1.744197151  0.72673106 -0.918423994
#>  -1.964590107 -2.28348075  2.621704368
#>   1.028025129  1.93259448  0.499181894
#>   0.549908786 -1.24079150 -0.289871254
#>  -2.099214118 -2.51462660  0.919867964
#>  -0.389655557 -0.24476894 -1.596275147
#>  -2.089725841 -3.09945989 -1.800882972
#>  -1.754535035 -0.69639415 -1.270956153
#>  -1.932657764 -0.20200539  2.017881345
#>  -0.290846473 -1.54705717 -0.828185565
#>   0.011308625 -1.57229011 -3.309097633
#>  -0.162691134 -2.57251983  0.005694352
#>  -1.518164548  0.69503377  0.347123616
#>   0.987906179 -0.21177210 -2.144164927
#>   0.856130253  1.70652697  0.725379255
#>  -0.639821594 -0.02614909  5.255050539
#>   1.794349088  4.09203722 -1.127407877
#>  -0.357504124 -4.49911111 -1.406544455
#>  -5.544715236 -1.64209714  2.762608270
#>  -0.872268035  2.96332732  0.154623743
#>   3.729953757 -1.65969500 -1.700585561
#>   1.248125883  0.57286541 -0.293028286
#>  -0.025788081 -0.02953902 -1.594008646
#>  -0.389090695 -0.60797199 -4.043294009
#>  -1.539614973 -2.73239187 -0.364772803
#>  -2.483362005 -0.20264493 -1.866035204
#>   3.472015243  0.22030849 -2.752989147
#>   0.894433101 -1.94907768  0.996999577
#>   1.757420362 -0.79802014 -1.530824085
#>  -1.942774130  0.95344661  1.631983484
#>   2.391253249 -2.24948779 -0.801017600
#>  -0.869982014 -0.42407842 -0.446844788
#>  -0.004483237  0.65261073  1.035622131
#>  -0.117373448 -0.68185758 -0.752457088
#>   1.981492832  0.50422644 -0.984415417
#>   0.588194989 -1.29795401  1.759823966
#>   1.455303879 -1.94506614 -3.317147047
#>  -1.970862810 -2.42787691 -0.662606830
#>  -0.540826257 -1.47063283 -2.749426602
#>  -2.137584973 -3.27209665  1.740648643
#>  -1.044875451 -0.55867780 -1.879800611
#>  -0.536571775  1.58985096  3.794070408
#>   1.632553261  2.87047038  1.184561284
#>   2.916050754 -0.78174558  0.467171138
#>   1.307125350  1.39972897  1.571466045
#>  -0.762402355 -2.43217981 -1.763434233
#>  -1.238616594 -1.76587615 -1.087068049
#>  -1.239195959  2.74951536 -1.469983310
#>  -1.663345275 -0.36982306  4.966654975
#>  -2.164444645 -3.38184985  4.985259140
#>  -3.090407192 -3.99134003  3.685231818
#>  -4.023295939 -1.71160205  2.986815759
#>  -1.363399834  3.88970311  0.433652163
#>   4.308614866 -1.60227423  0.172803034
#>   1.549764764 -0.75553158 -1.273221602
#>  -0.540372744 -3.63395027 -1.225791484
#>  -1.709038989  1.67632623  0.868566069
#>  -0.825855530  0.82564541  2.866879540
#>   1.950834316  2.37654786  1.375659291
#>   1.574618035  2.51293132  4.486589311
#>   1.630818661 -0.69883274 -5.511340577
#>  -1.447757198 -3.88789500  2.290642228
#>  -1.735237164  1.49506579 -0.594779728
#>   0.975572600 -3.58981892 -0.405227904
#>  -1.303753660 -2.50018488 -3.367728922
#>  -1.598202523 -3.31097870 -1.747931461
#>  -2.890533336 -1.40737479 -1.691006115
#>   2.214562899 -0.27981293  3.672233747
#>   2.425769905  2.08524809 -2.596656915
#>  -0.221227727  0.23138453  3.488043987
#>   0.442189230 -0.06461063 -0.673461121
#>   2.222917669  0.92011409 -4.263745515
#>  -2.755895057 -2.78350590  2.863728784
#>  -0.769712675  1.81567684  2.449582970
#>   2.499746428  2.03222524 -0.015148250
#>   1.582786025 -1.36549140 -2.341035559
#>  -1.867019576 -2.45585488  0.520729283
#>   0.972208495  1.33851695  0.839662851
#>   1.257739261  0.85701223 -0.267599499
#>   0.412855993  1.01968648 -1.000489728
#>  -1.536779967 -1.55355225  2.734978753
#>  -2.345284232 -1.45317520  5.329701253
#>  -1.932889326 -2.70828044  6.130083421
#>   1.664935052  4.31068951 -0.389230402
#>   3.623179321  0.49534793 -1.730328688
#>   0.549233340 -2.35526624  0.411302543
#>   0.988540832  0.54528445  0.270311523
#>  -1.325006701 -2.73237879 -0.809737551
#>  -0.881365238  1.23129188 -0.855469322
#>   0.549328346  0.13453583 -2.879850835
#>  -0.018733873 -3.89704313 -3.894341168
#>  -2.224130672 -2.50658514  2.913122759
#>  -0.922108987  1.76904675 -0.072604853
#>  -0.120882237  0.12069964  4.121205476
#>  -2.407690803  0.25859248  2.809268776
#>  -0.751710539 -1.51991267  4.156186195
#>  -3.507925796 -2.08882527  5.000547955
#>  -0.330793847  3.45171193  1.296031649
#>   3.232396214 -1.07649679  1.930206479
#>   2.055743063  1.16597920 -1.538638543
#>   0.634428647  1.95912665 -0.027336069
#>  -1.436469659 -0.76436513  1.476810439
#>   1.102766851  2.23872116 -1.056307097
#>   4.364955882  3.11705173 -2.777262677
#>  -0.776802034 -3.97873729  0.997759834
#>  -0.255973313  1.42330698 -3.519251523
#>  -2.605471118 -1.09566701  2.471088643
#>  -1.130762452 -1.47758496  1.012630793
#>  -2.442173560 -1.33717265  1.007446552
#>   0.655787508  4.75305729 -1.874057841
#>  -0.892081314 -1.13584492  3.711876865
#>   0.199025559  1.98123002 -2.497540840
#>   2.750226608 -1.64048736 -3.978456988
#>  -1.037669363 -2.15323922  1.240028242
#>  -3.092183270 -1.91032219  1.625306231
#>  -1.371003652 -1.19855846  0.579020609
#>  -1.024903289 -0.15635640 -0.319236979
#>  -0.490715389  1.09468327  1.188880142
#>   1.216704204 -1.68425788 -0.042219475
#>   0.574668775 -2.19725886 -3.376646388
#>  -1.181481700 -2.50008133 -4.796118514
#>  -1.247749816 -1.14447944  1.961594241
#>   1.066122214  1.12871064 -5.448083000
#>  -1.331480168 -3.55475243 -2.042392335
#>  -1.409332046 -1.99219619  0.127180654
#>  -2.461080407 -0.32582048  1.956500909
#>   0.056400467 -1.64896519  3.779951383
#>   2.927393830  1.57536213 -3.759201602
#>   2.142253253  2.28755397 -0.307564128
#>   0.242385023  0.57542001 -0.373429200
#>  -0.827997085 -2.70688002 -3.688359116
#>  -1.873102284 -1.77323609 -0.642621909
#>  -0.782293926 -2.64314881  2.210372689
#>  -3.020102997 -1.49674335  1.621357275
#>   0.128477633 -1.78525835  5.798854077
#>   3.357972353  1.80726017 -5.177958321
#>   2.843727368  2.06395255 -2.147217422
#>  -0.340483855 -2.04028786  0.017096396
#>   0.888733890  0.59683286 -1.591814921
#>  -0.348374351 -1.02679407 -0.899312784
#>  -0.566858380 -2.31583324 -1.249429825
#>  -0.100183682 -0.03129701  1.793671412
#>   0.861454224  2.36612067  0.385353375
#>   1.292424837  1.31043172  0.812196512
#>  -0.140597887  0.50017159  2.800037051
#>  -0.324712823  0.31891285  3.826245245
#>   1.195298041 -0.33236911 -1.624881611
#>   0.555021948 -2.45050929 -1.409532439
#>  -1.378860663 -3.98075310 -0.052254117
#>  -1.859318127 -1.87108542 -2.750539885
#>  -2.231380731 -0.93832241 -1.933836281
#>  -3.095894747 -1.29529705  1.564971282
#>  -0.082570451  1.30711063  1.060236774
#>   0.868855324 -3.03340536 -0.932159461
#>   0.678990709 -0.60737682 -0.447761729
#>   0.151932846  0.74404470 -0.014000847
#>  -0.181348554 -0.11593219 -0.518917664
#>   1.252851945  1.93700756  1.244326854
#>   0.768643017 -0.82851144 -1.973445803
#>  -1.059407971  0.20578983  0.191184919
#>   1.824064050  1.77368312 -0.738669611
#>   2.169177015 -1.09821734 -3.217941422
#>  -1.048835608 -3.01874565 -4.336754939
#>  -0.891058969 -0.24724137  2.546786854
#>   0.730523134  3.06297586  0.569114776
#>   1.508814563 -0.80396219 -0.663704377
#>   0.412823587 -1.69104654  2.352701856
#>   1.486239510  2.12955502 -1.071693712
#>   1.890197090 -1.29640718 -0.151329979
#>   0.959543263 -0.84662750 -1.655462012
#>  -0.385037605 -0.05603552 -0.547716856
#>  -0.307816985 -0.34882309 -1.449413799
#>  -0.388865426 -0.98445680  0.708066851
#>  -0.948423651  0.24770950  1.658294677
#>   0.766817437  2.17743018 -0.882658703
#>   1.756951966  0.99697367 -3.419421811
#>   2.637359803  1.01646623 -2.356416191
#>  -0.154779487  2.65312781 -0.951663311
#>   3.429045798 -0.47406714 -3.244127112
#>  -1.375434395  4.03430060 -1.631003731
#>   0.582945358  0.27919415  4.094413738
#>   1.518204206  3.20021002 -1.754821319
#>   3.544724725 -1.00258792 -1.143206864
#>   1.332956732 -4.86534210 -4.527121618
#>  -4.667974671  2.47125644  1.254278107
#>  -2.779538831  1.74430413  5.277578319
#>  -0.856318713  3.21354309  1.771270898
#>   0.546968311 -0.39081472 -0.051726787
#>   0.590733989  0.36389155  1.101830775
#>   1.424872046  1.24348208  2.927639421
#>  -0.456304711  0.34840245  1.688448992
#>  -0.565215682  0.83068747 -0.480061267
#>   1.502342682 -0.06150878 -1.469562389
#>  -0.854886882 -3.31610768 -4.749563221
#>  -2.196214125 -2.10864295  1.466011072
#>  -0.306713396  0.40075063 -1.538285631
#>  -2.577053063 -1.42854203  5.694034623
#>   1.693487200  1.52497357 -5.666472494
#>  -1.939185193 -2.74346412  2.226590890
#>  -0.877639490  0.68415553 -0.394907699
#>   2.064918020  3.97868750 -1.065328288
#>   0.534473727 -0.81112894 -0.067869512
#>  -0.238039533 -1.57716106  0.709291854
#>  -1.529745460  0.70084423 -0.858687682
#>   3.251175153 -1.61373279 -3.147176736
#>   1.944864452 -2.51503530 -5.466553757
#>  -0.792547218 -4.19147496 -0.304423733
#>  -2.629786120 -0.15467195  3.145976288
#>   1.491501298  1.71569655 -1.586426494
#>   0.710695181 -0.79433129  0.820466688
#>   1.062896669  0.72085154 -0.034257895
#>  -0.867212168  0.47063916  3.935372954
#>  -0.712331055 -1.66074760  2.817577592
#>  -1.152684977 -0.75384499  3.305140893
#>   0.230581684  0.87466920 -1.706510897
#>   1.245010750  1.22388652 -1.249798446
#>   0.624193421 -0.70127379 -2.482087134
#>  -0.256182527 -2.39782506 -3.686658292
#>  -0.795618713 -2.69630334 -3.152244284
#>   0.387550447  0.54789868  2.259787416
#>   0.787530628 -2.40693017 -2.532300864
#>  -2.183085528 -2.43396072 -0.261758400
#>  -0.449850008  0.38586407 -2.717444093
#>   2.209231946 -0.52385389 -5.730203883
#>  -2.390852743 -2.69215318  1.119831529
#>  -2.461345665 -3.88280642  0.131728834
#>  -0.557897718  0.49446391 -2.130777999
#>  -2.573782151 -2.10951424  0.561215581
#>  -2.370681956 -0.34924417  2.313748864
#>  -0.726529710  2.28370552  1.234528702
#>   0.349439656  0.40474765  2.150429590
#>   1.428578950  2.69184372 -0.091533856
#>   0.418434116  1.52726260  0.488575460
#>   1.442117994  2.31442082  0.036811644
#>   1.029780701 -3.88609135 -2.860767779
#>  -4.237700047  1.14785824  1.763478796
#>  -3.105713458  3.80050079  9.841562041
#>   2.362370943 -4.03245645  0.666666537
#>   1.222309993  1.63785480 -3.365728618
#>  -1.065005907 -0.22256034  4.815322204
#>   0.825528135  2.80268276 -1.929112477
#>  -1.021508454 -0.10287524  1.254691736
#>   0.300705912 -0.22265476  0.348181389
#>  -2.168940484 -1.69660446  1.656618768
#>  -0.897078640  0.94726403 -2.269443607
#>  -3.054226779 -1.41158117  2.903933815
#>  -1.512853716  2.71479748  3.496918353
#>   1.853094363 -1.29157042  0.204054596
#>  -2.053113999  1.43622587 -0.572493648
#>  -0.764634927 -0.53190344  2.114358677
#>   0.857343608 -0.01368488 -0.271643504
#>   0.702156978 -0.70096234 -1.491381586
#>  -0.448101379  0.27472110 -2.340572460
#>  -0.496676513 -1.38271093 -1.391759674
#>  -1.919034966 -1.91948909  0.817708541
#>  -0.616500037  2.57128167  0.257812418
#>   2.224983549 -1.00097963 -2.891797688
#>  -0.289245760 -3.29431181 -1.274003487
#>  -2.381378828 -2.30701540  0.210022682
#>  -0.626156932  0.54926659 -2.238295442
#>   2.158000044  0.19964096  0.761380841
#>  -0.876945053 -0.72698113  1.701956546
#>  -0.213402209  2.15741822  2.661178317
#>   1.087703773  1.85823291 -2.762157786
#>  -3.556960659 -4.75267570  1.311545444
#>  -4.685282464 -1.93116258  3.503343385
#>  -0.558036494  3.58737887  0.691806243
#>   1.061090293  0.67164320  3.156380013
#>   1.895805547  0.26104398 -3.162764386
#>  -1.791256556 -3.05701072 -2.931674429
#>  -2.814807108 -2.63642341  0.401911541
#>   0.356112507  2.46007993 -1.426305719
#>  -2.309674160 -1.86910053 -0.465147509
#>  -1.781199514  0.75012705  0.587605376
#>   1.361174594  1.64959550  3.032695847
#>   0.740444836  0.76666643  0.767915062
#>   0.651482591 -1.74374720 -1.170826604
#>  -0.103603726 -1.14814979 -2.028611999
#>  -1.793901727  0.97357086  2.183554901
#>   2.101598868 -0.88023779  2.777274057
#>  -2.626032473 -0.01324652  0.454618840
#>  -2.407092256  1.39068272  3.649331563
#>  -0.320962903  1.45487698  2.516294577
#>  -0.214952338  0.81549349  0.141033865
#>   0.325124351 -1.91794772  1.057041662
#>  -0.432580944 -0.25711721 -1.376616703
#>   0.489313253 -0.01005865 -2.515187704
#>  -0.270385499 -1.00914679 -0.487157274
#>   0.517922200  2.09719992  0.745027914
#>   1.313551171  1.23081108  2.134563623
#>   1.843558710  0.50959706 -0.268843101
#>   1.293800768 -1.59963703 -0.646105868
#>   1.388563052 -1.32682119 -1.760621856
#>   0.296827590  0.27608658 -2.267243870
#>  -0.597233410 -1.71096427 -0.702858570
#>   1.419365454  1.05708496  1.578198140
#>   1.753343912  2.27047340  3.402024754
#>   0.172666967  1.09533678  3.180573718
#>   3.252763995  5.74950434  0.543857737
#>   1.793774552 -0.07198775  2.931199130
#>   3.014417194  1.69813879 -1.142778847
#>   1.493341750 -1.04005276  1.269242438
#>   0.144496773  0.71387445  1.849768430
#>   0.943869667 -0.09575402 -1.304224554
#>  -0.330418738 -2.34364575 -0.516193252
#>  -0.561100962 -0.38120795 -0.093334172
#>  -0.420018742 -0.65905733 -0.873994741
#>  -1.141375145 -1.33476811 -0.893326127
#>  -1.160813891 -1.49140628  0.788893897
#>   0.263476603  2.35232665 -1.919798846
#>  -0.184570325 -1.67882020  2.015147706
#>  -1.448180661 -0.16750703  1.261864639
#>   0.663699157  1.34717110  0.107588472
#>  -0.041814175 -0.56962126 -1.086506178
#>  -0.752655372  1.95155544  2.214635551
#>   4.134616285  0.84082161 -0.283009369
#>   2.161367162 -0.65587592  0.539490428
#>  -2.122785846  3.33939789 -0.927503086
#>  -1.576333982 -0.08985839  2.780964022
#>   1.792509145  1.52347583  0.758924572
#>  -0.766864497 -1.83565022  1.069193600
#>  -0.793180586  0.34505010 -2.365919328
#>  -1.929679150 -2.45666668  1.049961485
#>  -1.845259469 -1.22803698 -2.173268096
#>  -2.559467257 -3.66131333 -0.842866115
#>  -1.173695754 -0.65176697 -2.585348811
#>  -1.072172979 -3.01086785 -1.305592683
#>  -1.254112770 -2.67362718  0.177144909
#>  -0.264396036  1.80254246  1.490371610
#>  -0.528579142 -0.13842512 -2.909002117
#>  -0.203238925  1.39800698  3.985950995
#>   1.466381238  1.74409585 -0.930094420
#>   1.310423069 -1.73663093 -1.729654592
#>   0.204303070 -2.37764548 -1.750175579
#>   0.642613676  1.49307877 -2.325213180
#>  -1.913420099 -1.92192421  4.863984613
#>   1.794744780  3.50361959 -2.922537161
#>   2.483170749  1.71138394  0.636686399
#>   0.597500458 -1.09215787  1.445638504
#>  -0.303985450 -3.11501425  1.384633066
#>   0.496882195  1.77895063 -0.550484972
#>   0.292424834  1.42141144 -1.681848924
#>  -0.349720719 -1.75616949 -3.114982553
#>   2.087546032  1.19373937  2.042812627
#>  -0.552097197  1.03400732 -1.089400413
#>  -1.757357381  1.31253061  3.080767949
#>   0.560742662  2.18694854  0.221589953
#>   1.180706154  0.24211294  0.085844875
#>   0.160246138 -0.24799298  0.405361025
#>   1.187193947  0.18967713 -3.494903218
#>  -0.808039714 -2.25027851 -0.423776761
#>  -0.273761812  1.35081753  1.775207691
#>   0.344695567  0.16419342 -1.690945080
#>   0.607814195 -0.54830142 -1.355304650
#>   0.457740344 -0.57484741  0.627761286
#>   1.746376021  0.60135009  0.761963735
#>   0.848525929  1.03150242 -0.197207439
#>  -0.707158682 -0.54441820  2.406180810
#>  -0.063814084  0.73933929 -2.027313389
#>  -0.382003817 -0.23645861  1.199956457
#>  -0.459061440 -2.03204700 -1.137611916
#>  -0.050089232 -0.67965231  3.082164342
#>   1.519835717  1.75749972  1.639140492
#>   2.493168268  0.40758362  0.417848182
#>   0.060903193 -0.47673587  0.674697849
#>  -0.274738724 -0.44996907  3.215969894
#>   1.052656714  1.35103175  0.248657818
#>   1.833062824  1.50723754  2.229505905
#>  -0.978354702 -0.03426939 -0.985887992
#>  -1.298410901  0.84109536  0.346936495
#>  -1.250579780  0.50622378  1.483792505
#>   2.125857912 -1.72806574  3.039120615
#>  -3.732451873 -1.08028571  3.729603492
#>  -0.563398650  4.18230066  3.232587160
#>   2.423651565  3.08690355 -0.588940019
#>  -0.228631704 -1.01725600 -1.876340287
#>  -1.221688201  0.52792669  3.799786585
#>   0.073169483 -1.42952409  3.741366123
#>   0.106322847  0.51812916 -3.606201678
#>   1.977807404  0.05646004 -4.084299836
#>  -1.622997566 -3.52139782 -0.240614213
#>  -1.828080363 -0.82513342  3.216493988
#>   1.696548669  2.45899409  0.626801825
#>   1.169601829  0.71707907  2.021098446
#>   3.240319658  0.81598026  0.683075197
#>  -2.264031163  0.67125457 -2.479477157
#>  -0.925796010 -0.99917081  1.246759630
#>   0.262005689  1.89278611 -3.966578702
#>  -0.883465381 -3.16376616  1.204642736
#>  -3.309436045  1.08350168  1.593031477
#>  -0.154432873  1.78338793  1.770309008
#>   0.631397941  0.20848759  1.654621864
#>  -0.689302846 -0.38692014 -1.176900511
#>  -2.533002264 -1.59096205 -1.549319986
#>   1.322922977 -2.21622348  2.539491001
#>   2.730503819 -0.13278800 -5.950012195
#>   0.384146238  3.46919450 -3.381380390
#>  -2.661443496 -3.40153927  4.886902650
#>   1.261359441  4.02106838 -0.423189032
#>   5.294332000  0.46822753 -1.312979603
#>   2.541195194 -2.16751909 -3.084254560
#>  -0.155730950 -2.52684155 -2.159211151
#>  -1.707820667 -0.38030728  0.657340301
#>  -0.479823579 -1.51102941  1.260318657
#>   1.032812026  2.68870641  1.100645285
#>   2.391758319 -0.87762502  0.600349453
#>  -0.046144849  3.39819623 -0.332274690
#>   3.826329563  0.58589894 -1.813271342
#>  -1.553970721  0.78612874  1.150949820
#>   1.905943204  1.60828761 -1.724136877
#>  -2.603799255 -2.11278116 -0.145387161
#>  -2.744684041 -1.47933295  2.828274356
#>  -1.133177912  1.06706498  1.754263003
#>   0.956629417  1.80349028  0.226740224
#>   0.467895982  0.13391107 -0.436538567
#>   0.048352125  1.27079717 -1.741842579
#>  -1.785474388 -2.69487282 -0.128415656
#>  -0.359599148  0.02064478 -1.738455467
#>  -1.012519153 -0.36762400 -0.178588131
#>   0.420828177 -0.30640537 -0.448959531
#>   0.272182286 -1.62731314 -2.086359666
#>  -0.449993348 -3.75015858 -3.473409422
#>  -0.588964785 -1.37358648 -1.914651376
#>   0.138797715  2.13319768  1.266617901
#>   0.923032019  1.10846554  0.522056656
#>  -0.096309277  0.65536694 -0.087439643
#>  -1.511732469 -1.26004970 -0.280881483
#>   0.249020399  1.75743528 -1.599490033
#>   0.222256261 -0.57195164 -0.170339632
#>  -1.440500129 -1.72122100  2.014155425
#>   0.356530555  0.03416265 -1.643748167
#>  -1.698425903 -0.61997163  3.666480897
#>   0.507172436  4.09211415  3.560565211
#>   3.658211139  0.21168928 -1.145912769
#>   1.899409975 -0.87826558 -1.717235448
#>   0.108769210  0.16067307 -0.570794845
#>   0.356998727  1.49056399  3.844340138
#>   1.459472740  0.92326090 -0.360732755
#>   0.225933944  0.27604226 -2.108275022
#>   1.674187519  0.52785486 -0.280991725
#>  -0.106434966  0.73618024  2.365676188
#>  -0.735635264  3.17643950 -0.508382099
#>  -1.538046775 -0.24852871  1.603499891
#>   1.370327787  2.29927490 -1.066189767
#>   0.907550084 -1.14136949  0.379332278
#>   0.866804428 -0.43043734 -3.384707533
#>   0.693868716 -2.74927035 -1.922925082
#>  -1.530763334 -1.08472106 -2.661483568
#>  -2.182446360 -0.88780385  0.526627703
#>  -1.184495356  1.23712577  2.276641071
#>   1.820957543  0.87401488  2.228529771
#>   1.947222232 -3.19832668 -2.287545953
#>   1.179780089  0.32328916 -4.523407371
#>   1.163465145 -1.73274162 -1.358644358
#>  -0.553548378  0.83466052  1.603913157
#>  -0.260004697  2.24979395  0.598253898
#>   0.239910624  0.62941421  2.008961676
#>   1.448411760  2.39198139  1.310286193
#>   2.757135324  1.33571396 -0.622667466
#>  -0.445179903 -1.81684173 -3.489139572
#>  -0.764139784 -2.95116850 -1.018762871
#>  -1.293507299 -1.31664744  2.149260311
#>  -1.402543719 -0.56879773  0.689256498
#>   0.378921804 -0.35620968 -1.135473338
#>   0.346755303  1.97158881 -0.752160610
#>   0.820969202  0.85715063  1.044096711
#>   1.047347700 -0.41719305 -1.840965841
#>  -1.007653360 -4.14761619 -0.333182589
#>   0.423363159  1.50600069  1.467890587
#>   0.415880113 -2.20951575 -1.440388845
#>  -2.327393965  1.70931463  1.584232392
#>   0.893703195  1.68509509  3.070839965
#>   1.151971885  0.50893337  0.106662590
#>   0.007828798  1.07292444  0.631632710
#>   0.107658320 -0.97427651 -1.814142977
#>  -2.725876495  1.01946257 -1.138455995
#>  -2.537954798  2.05858038  7.464168933
#>   0.329135296 -2.02147013  4.418163093
#>   3.550039676  2.79053436 -2.861136352
#>   0.709744372 -2.13739363  0.856250671
#>  -3.564904731 -0.30224994 -1.936162721
#>  -3.005877060 -0.44489060  4.085967669
#>  -0.706011190 -3.03781574  1.242294038
#>  -0.587611438 -0.48752785  0.274915294
#>   0.392507716  0.58892565 -3.012405564
#>  -2.508487262 -2.34221747  3.503254487
#>  -0.023107568  4.11547436  0.614281457
#>  -0.311599378 -0.35480635  4.425038953