Simulate from multivariate MixAR models

Simulate data from multivariate MixAR models under the assumptions of multivariate Gaussian innovarion

mixVAR_sim(model, n, init, nskip = 100, flag = FALSE)

Arguments

model
model from which to simulate, an object inheriting from class MixVAR.

n
size of simulated multivariate series.

init
initial values, a numeric matrix. If missing, a matrix of 0 values is generated.

nskip
number of burn-in values.

flag
if TRUE returns also the regimes.

Details

mixVAR_sim simulates a series of length nskip + n and returns the last n values. init provides initial values for the algorithm. Each row is considered as a time point. The number of rows must be at least equal to the maximal AR order.

Value

a numeric matrix with n rows.

See also

Examples

AR <- list()
AR[[1]] <- array(c(0.5,-0.3,-0.6,0,0,0.5,0.4,0.5,-0.3), dim = c(3,3,1))
AR[[2]] <- array(c(-0.5,0.3,0,1,0,-0.5,-0.4,-0.2, 0.5), dim = c(3,3,1))

prob <- c(0.75, 0.25)
shift <- cbind(c(0,0,0), c(0,0,0))

Sigma1 <- cbind(c(1, 0.5, -0.4), c(0.5, 2, 0.8), c(-0.4, 0.8, 4))
Sigma2 <- cbind(c(1,0.2, 0), c(0.2, 2, -0.15), c(0, -0.15, 4))
Sigma <- array(c(Sigma1, Sigma2), dim = c(3,3,2))

m <- new("MixVARGaussian", prob=prob, vcov=Sigma, arcoef=AR, shift=shift)
mixVAR_sim(m, n=500, init=matrix(rep(0,3), ncol=3), nskip=100, flag=FALSE)

#>          [,1]         [,2]        [,3]
#>  -1.588611753 -0.537883397 -2.66468079
#>  -1.537259752 -1.102505690  3.06454501
#>   0.924844611  2.110519962 -0.78397834
#>   0.020516152 -0.755478364 -1.35084266
#>   0.140918440 -1.252806248 -1.65964454
#>  -0.199891993  1.187465826 -3.17239253
#>  -1.595615547 -1.594606469  0.68267173
#>  -0.866077906  1.305414059  0.55488188
#>  -0.755075889  1.205375844  1.60742793
#>   0.737659696 -0.824815726 -1.38361985
#>  -0.065819121  0.226944227 -1.21722096
#>  -0.217878572  1.206391570  0.17317976
#>   0.196518437  2.620234393  2.38844581
#>   0.486242356  3.339167842 -2.19148845
#>  -0.740341447 -2.878786215  0.98886148
#>  -0.059950808  0.798735331 -3.38122247
#>  -3.936249794 -2.627169805  0.37796616
#>  -1.933835307  0.451394337  2.25758391
#>   0.593287054 -0.134023213 -1.11282286
#>  -0.258084142 -0.333248773 -0.14722043
#>   1.340819340 -2.713651284  1.69330574
#>   1.595379945  1.721311316 -2.63109931
#>  -0.719559410  1.379880889 -2.63469983
#>  -2.524139475 -2.217252595  2.01249933
#>   2.093966622  2.343710496 -2.47088220
#>   4.144656149  0.698268776 -4.04472355
#>  -1.200393980 -3.996016272  2.93015123
#>   1.589815000  3.082206318 -1.82418572
#>   0.439630689 -0.460876621  1.66439353
#>  -1.112879795  1.165123140  0.68922114
#>   1.627357140 -1.786639568  2.96903078
#>   2.954071456 -1.223765439 -3.85735952
#>   0.455194748 -1.835431530  0.31284227
#>  -0.400603471 -1.662471649 -0.91959744
#>  -2.009096329  0.352480177  1.47380851
#>  -1.054043194  0.278923594  0.06814031
#>  -1.243378556  1.126088041  3.50097717
#>   1.806813385 -2.617968232  0.36701347
#>  -3.517362700  0.593549046  0.59990766
#>  -0.246317795 -1.710609952 -2.35046221
#>  -1.206686467 -2.272997114 -0.44918174
#>  -1.530672694 -0.688637529 -1.01145976
#>   0.307507805  1.185795162  0.38986655
#>  -0.402387820  0.899149651 -2.09788087
#>  -2.203872503 -2.509407107 -0.61262037
#>  -1.901061591  1.886663276  0.50317191
#>  -1.312441958  1.627979513  2.60303622
#>  -0.999484373  0.776554819  1.92645914
#>   2.378859819  4.094106137  2.15832993
#>   1.755444567 -2.188912287 -1.90207422
#>   0.013968982 -0.839675194 -1.30377031
#>  -0.094313179  1.977700596  1.17931653
#>   0.041311977 -0.011306305  0.81700764
#>  -1.055949273  1.724697026 -0.22198288
#>  -1.471762971 -0.881226192  3.27847304
#>  -2.768150328 -1.878539175  2.72380108
#>  -0.135710931  2.183253522 -2.61986220
#>   2.727840706 -1.920056266 -4.01095105
#>   0.907457102 -2.190288808 -1.34046049
#>  -0.949459725 -2.176010715 -3.24164929
#>  -1.589865299 -0.198652788  1.22859100
#>  -0.829407482  1.804222273  1.48747743
#>   1.072015064  0.288219848 -1.62080777
#>  -0.366394341 -0.346961492  3.16013996
#>   1.125544245  2.239344391 -0.94902253
#>   2.599397042  1.464583821  0.26747565
#>  -0.070732021  0.680153483 -0.37130193
#>  -0.599633564 -0.357731287 -0.77438186
#>  -0.772928483  0.990135335  1.06790307
#>  -0.180918754  2.333819719  0.66352339
#>   1.796913738 -1.555190023 -0.84221253
#>  -2.247120454  0.771816390  1.24347101
#>   2.329869407 -0.429178362  2.20716665
#>   2.450730869 -3.462626078 -3.88928438
#>   0.108040647 -4.324858878 -4.30344558
#>  -1.236773341 -3.158825089 -4.51430218
#>  -1.808458609 -2.734788468  0.93041722
#>   0.184774183  3.240347026  0.14269282
#>   1.115434360  1.633999376  0.84346302
#>  -0.868399855  0.231725594  2.81116675
#>  -1.775928392  0.491246069  1.12322739
#>  -0.585241750  0.896635347  2.92276800
#>   2.757712609  3.880907004 -1.82103571
#>   1.445933806 -2.324214035 -1.32538745
#>   0.439842439 -0.833560204 -0.13100589
#>   2.236742784  1.466097531 -2.03811926
#>   1.420976043  3.032769365  0.53417069
#>  -0.326300728  0.689045692  0.31280822
#>  -0.844410452 -1.479585838 -0.34768313
#>   0.409969743  2.243441988  0.39676693
#>   1.791302370 -0.289922543  0.99261457
#>   0.152947918  0.307286383  0.73424377
#>   1.667325687  0.849032510 -2.51916996
#>  -0.904243765 -3.298507450 -2.72717482
#>  -0.841113438 -0.014646180 -4.82990243
#>  -1.551179059 -2.457542731  0.25611391
#>   0.201152412  1.685315211 -0.14271104
#>   0.664818854 -0.910971275 -4.11779877
#>   0.573302940 -2.225768249 -1.44337264
#>  -4.124903665 -0.300731185 -0.69052873
#>  -3.270835331  4.058695880  5.07173825
#>   2.230563228 -0.569048396  2.54391417
#>   2.768936093  1.323037552 -2.09157496
#>   1.005128890 -3.312987575 -2.01098670
#>  -1.325329044 -1.729572837 -1.25149895
#>  -0.687974247  0.997805648  3.20405359
#>   1.630556846  3.944745692  2.34189384
#>   1.289756959  1.690355969  1.46770052
#>   1.046220915 -0.718991291  0.55027300
#>   0.578824731 -0.280978800  0.57418389
#>   2.080482534  0.890435202 -5.76196624
#>  -1.793829392 -2.611715593 -0.17541220
#>  -1.220959277 -0.259752000  3.42486818
#>   0.010148924  1.921359878  0.35378767
#>   1.282687844  0.296717771  1.18052375
#>  -0.062151832  0.815934273 -0.19923324
#>   0.145946454 -0.076628262  3.63710179
#>  -1.733848034 -0.519666857  1.64005587
#>   0.393063310  0.531010960  0.36351423
#>   0.570717517 -1.456052825 -0.02901766
#>  -0.076331839  1.814377803  0.78162875
#>  -1.203970229  1.331555196  1.64721797
#>  -0.366230960  0.249676406 -1.62393190
#>  -0.818352318 -2.897169070  0.08538432
#>  -0.052394983  0.722120126  0.66640184
#>   0.722935233  3.151141218 -1.84792451
#>   0.895398982 -3.270386996 -1.07745436
#>  -1.431469866 -2.733234437  0.84880969
#>   1.028098547  2.823073837 -1.78586796
#>  -0.024914833  0.618397164  3.69414407
#>   2.220978945  1.910584495 -3.13490695
#>   2.347042593  1.095943416 -3.82324409
#>   2.964440950  0.599298352 -1.94116591
#>   0.495585678  2.787546562 -0.58143311
#>   2.089212624  0.505523197  0.10828634
#>   1.682019232 -0.176860490 -5.21568833
#>  -1.290302112 -3.137039272  0.02029045
#>  -1.101624957 -2.238767884 -0.62036789
#>  -0.463727175 -1.369552099 -6.06218324
#>  -3.013293553 -2.842425487  1.36486412
#>  -1.329241770  1.660241137 -1.64885535
#>  -0.610963183  1.108239244  4.11369317
#>   2.029786822  1.935025600 -0.57500276
#>   2.067770290  0.316316417  0.56689245
#>   2.340491378 -0.895654202 -1.55816640
#>  -0.166216722  0.763509677 -0.04874295
#>   0.020414771  1.719275229  1.09468021
#>   1.281088807 -1.150069644 -1.11895774
#>  -1.351574703 -0.532460583  1.01571294
#>  -1.512486047  1.237474920  1.19910501
#>  -0.327099090  1.604368942  4.38823462
#>   0.461509402  1.010119859  2.23246060
#>   0.330455664 -1.147589358  6.01602984
#>   2.339748206  2.307556544 -1.93514498
#>   1.098424845 -0.809721094 -1.36265446
#>   0.311179652  0.998093319 -1.19636040
#>  -0.429744277 -0.087302670 -0.31995250
#>   0.830383156  0.833244339  0.41605820
#>  -1.360942731 -0.545766772  3.87234909
#>  -1.645906559 -0.242426972  0.33086253
#>  -2.280210611  1.627210559  4.05889554
#>  -0.855938426  3.600828686  2.07029208
#>   0.608973764  3.357928029  0.19786156
#>   0.593795817  0.536762371 -0.70213757
#>   0.860472136 -0.123855391  2.35362167
#>  -1.581473721 -1.192817814 -1.40931769
#>  -2.395194077 -1.252443551 -3.01703703
#>  -0.239339354  2.532647141 -1.32712935
#>   1.015598857  2.015233371  2.24439614
#>   1.348554361  1.257593572  1.11808310
#>   0.767048419 -2.598187259  0.11468112
#>  -2.430459487  0.008275402  1.81607471
#>   1.376003527 -1.294147737  3.98510500
#>   2.377368468 -0.552888111 -3.50702393
#>   0.238751027  1.751716338 -4.22683571
#>  -3.503494010 -2.542094711  3.63672620
#>  -0.980248056  3.369232038  1.09339004
#>   0.980763128  1.933471369  2.42475345
#>  -0.294747901  0.073059493  1.46585004
#>   0.615900890 -0.227302318  0.11034870
#>   0.559133298  1.575791315 -0.25354085
#>  -0.118173627 -0.960586944 -1.21184548
#>   0.829845661 -0.195627820 -1.39923636
#>   0.502133031  0.217355011 -0.85839461
#>  -0.087626172 -1.231040303 -3.86250448
#>   0.515890098  0.830719829 -1.84136694
#>  -1.142077828  0.037013801  0.17779379
#>   0.210138599  3.215444170 -0.01438602
#>  -0.462976630 -0.084639607  1.06389962
#>  -0.220204050  1.176782622  0.54897067
#>  -0.117611061 -2.239832074  0.94172614
#>  -0.901521432 -0.690889306  1.00607123
#>  -0.220426143  0.243316656  1.27746959
#>   2.038513397  2.106005847 -2.08813114
#>  -0.112381035 -1.701243256 -2.06999735
#>  -1.244152681 -1.125439725 -0.23569339
#>  -0.598152491  0.780653555  2.02734509
#>   0.288080669  0.011485223 -0.29716987
#>  -0.217522821  0.119122755  2.80493683
#>  -0.179331432  1.897059716  3.49239454
#>   2.825382313  0.164679815 -2.01087489
#>   0.061038295 -2.501032837 -1.74453778
#>  -1.482231682 -1.351055107 -0.62036398
#>  -1.516397976 -0.658895577 -0.80244273
#>   1.683424062  0.879994085  0.37244313
#>   0.163618133  3.614973399  2.12016284
#>   1.712780033 -1.355969819 -0.03117424
#>   1.576391317  0.332498246 -0.51410844
#>   0.301959389  0.021822457 -0.90485017
#>  -0.546711702  0.250883389 -3.32895015
#>  -1.654556477 -0.502311018  1.28754625
#>   1.410238267  2.414376444 -1.37332987
#>   0.059036011 -1.040601612  1.31761783
#>  -0.052349319 -0.393419051 -1.54174279
#>   0.586847059  0.847413321 -0.75846499
#>  -0.069867255 -2.105298489 -1.45898366
#>  -1.197265943 -3.158049735 -4.77748990
#>   0.149625163  1.269300264 -2.48966901
#>   3.292460524  3.289733822 -7.05679858
#>   5.714712956  4.220641875 -4.92176956
#>   0.268845265 -6.027775102 -1.77566789
#>  -1.573076978 -4.167660146 -4.34717410
#>  -1.995748752 -2.119582132  2.91856576
#>   0.073084864  2.317476777  0.23818605
#>  -0.882833883 -1.287926921  0.15864639
#>   1.437979419  1.315599438 -0.51840796
#>  -0.315546424  0.301642051 -0.12965659
#>  -0.613689684 -1.508525505 -1.65857930
#>  -1.253486123  0.015772981  2.14990543
#>  -0.915615943 -0.467435804 -3.92592600
#>  -1.796665709 -0.616443021  2.15765143
#>   0.691100766 -0.019358299  0.18890625
#>   1.822927683  0.974076752  0.54765441
#>   0.377866836 -1.133034437 -3.10689898
#>  -0.737022682  0.075959698 -0.41449037
#>  -0.043387457  0.709044514 -2.35093691
#>  -2.538296548 -2.676906123  1.66283245
#>  -0.046564300  3.015000978 -2.45789672
#>  -2.165572240 -3.797509062  5.01922552
#>   1.857694122  2.466984867 -2.93403347
#>   2.726614341  3.664620515 -2.70291302
#>   0.732947033 -0.941644781  2.48200962
#>  -3.905837754  0.669572523 -0.17058408
#>   2.499224020 -3.869977485  0.09322791
#>   0.078860166 -4.343659856 -7.15830107
#>  -4.109580002 -3.097743481  0.90103753
#>  -1.187567093  1.156783234  0.76680995
#>   0.260939746  0.441883061 -1.21377858
#>   0.800217703  2.369061477  4.66310869
#>   3.099226231  2.008783294 -1.13284100
#>   0.750535987 -1.940393406 -0.63514042
#>  -4.026225230  0.193763322  1.12030022
#>  -2.827511228 -0.953384852  1.17407333
#>  -2.980233378  0.987267410  2.36707468
#>   0.558317632 -1.420000877  0.35721732
#>  -1.525734544 -1.337241783  0.79786970
#>  -2.108690073 -1.740824789 -0.39059349
#>  -0.162489341  0.596981784  1.36986078
#>   1.295564162  0.108569821 -2.46248538
#>  -0.481343685 -1.712849012  2.75426857
#>   0.183607023  1.145343065 -0.03979404
#>  -0.642905494 -0.128991816  1.55568723
#>  -0.932986588 -2.261103504  1.52050693
#>  -2.849211030 -3.469301645  2.51838508
#>  -0.596491683 -0.291318788 -2.26028791
#>  -1.585043430 -1.618085878  1.87043211
#>  -1.708366910 -0.375487640 -1.30173387
#>   0.135131083 -0.272894354 -1.05700748
#>   0.983008218 -0.393391214  0.89418346
#>   1.818681499  1.297841598  1.44498250
#>  -0.847070697  2.526902077 -3.64266767
#>  -2.989107389 -2.539381332  3.33122743
#>  -0.547511132  3.358248393  0.57452602
#>   4.200146954  1.931853768 -4.30615341
#>   1.372949334 -3.753868934 -2.06833801
#>  -1.997016095 -3.096682096 -2.60773380
#>  -4.067641428 -1.519841899  1.65363247
#>  -2.410666560  1.405191395 -1.27498303
#>   3.140706376  0.886147806 -5.73824093
#>   1.074055156 -4.647080931  1.49006617
#>  -3.787473055 -1.105128905  1.76206826
#>   1.032800930 -2.515207995 -0.44295347
#>   2.301078306  2.202081251 -1.99907827
#>   2.992096848  1.433201957  0.30407055
#>  -1.049749480 -0.242533742  2.26745250
#>   1.051065241  2.451873732 -0.12068190
#>   0.363298703 -0.720826571  2.72695089
#>  -2.151956081  2.321369182  2.80908976
#>   1.344984254  4.292329823  1.73045995
#>  -0.566079733 -0.635352218  2.11681173
#>   1.636871398  2.001373762  0.91152566
#>   2.209792523  1.780299240  1.33360650
#>   0.035723251  1.848425590 -1.27687649
#>  -0.350182141 -1.148696656  2.04975531
#>   1.097332595  1.107512823 -1.02949451
#>   0.133256633 -2.041815593  1.87368917
#>  -2.544203610  0.469118257  1.59517906
#>  -2.224802547  2.089371706  1.24717153
#>   3.487782630  2.075168297 -2.82116532
#>   1.839022130 -2.663514843 -0.62645482
#>  -3.905646293 -0.020033543 -1.25990219
#>  -2.197285116  1.173487878  2.44065144
#>  -1.834883278  1.055633127  2.14125183
#>  -0.707277331  0.352848474 -1.36242647
#>   1.789176946 -0.973480473  5.50685135
#>   4.031564355  2.223379527 -3.40668107
#>   0.028037608 -2.690449461 -0.08895572
#>  -2.202179380 -2.833003584 -0.33655849
#>  -3.936791227 -1.395440164  4.57165860
#>  -2.886405722 -1.886741949  6.10759064
#>   1.939161581  5.561804820 -2.71673284
#>   6.325979413  1.376561809 -1.44697641
#>   4.828937892 -1.616079755 -0.86427774
#>   1.685139157 -2.604639663 -4.56654435
#>  -2.262100430 -3.010728076 -0.79132275
#>  -0.661338766 -0.073751451 -1.88799842
#>  -0.854383028  1.863271183  2.29666708
#>   0.720304321  3.365305051  4.52115623
#>   3.724431282  1.827583263 -0.63674463
#>  -0.158053053  0.911021782 -0.32141650
#>  -0.900511021 -1.883712446 -0.04194847
#>   1.416134309  3.026949744 -0.95020180
#>   1.111326439 -0.916538300  0.54745183
#>  -2.492078540 -2.138731959  4.40391877
#>   0.925114671  0.587610229 -2.15444018
#>   1.088199275 -0.801571928  0.55042825
#>  -0.636482683  1.657017478  1.88830407
#>   0.910300828  3.682162157  4.32897967
#>   2.356108017  1.661478034 -0.06724440
#>   1.230534141 -1.429715885  0.48414383
#>   0.471350802  2.011406940 -5.75515206
#>   4.245802593  1.199584508 -2.78583689
#>   1.191633928  0.500452312 -0.73400591
#>   0.315810126  2.061756973  0.72853733
#>   0.620418727  0.019978071  0.08839217
#>  -0.879901031  0.521495890  0.86418131
#>  -1.340711835  0.518979307  3.19188941
#>   1.423148308  0.816145879  1.18991702
#>  -0.213785278  0.786112470 -1.57402790
#>  -1.295690203 -3.780767032 -0.94990055
#>  -2.951512242  1.661619342 -0.75911140
#>  -1.256533595  0.358969979  0.10776559
#>  -0.095138074  2.900497463  3.11105740
#>   2.084996989  1.790973742  2.23937263
#>  -2.614231993 -1.780128483 -0.07314920
#>  -1.225148939  1.662379319  2.24005787
#>  -0.215131188  0.633086965  0.15478073
#>  -0.750639785  0.059924354  2.19158525
#>   0.952675355  0.652123526 -1.72867154
#>   0.747368138  0.275690251  1.25665498
#>  -1.625785064 -1.301577886  0.65201471
#>  -0.143567935 -1.472500614  0.58888229
#>   1.358969187 -0.312345306  2.13613890
#>   2.162694398  2.043088350 -2.44103736
#>  -0.944573737 -1.476230473  1.56348655
#>  -0.660198113  0.087346655  1.77874129
#>  -0.479792636  2.592327029 -0.70557476
#>   1.419906823  2.900674349 -0.98983395
#>  -0.675846468 -1.881101484 -0.60404603
#>  -0.977911854  0.187836139 -1.84748426
#>  -1.235903590 -1.338491471  2.27950224
#>   0.987487953  1.405914005 -2.27609077
#>   1.270658562 -0.753481420 -0.13029963
#>   1.842750029  0.019143054 -0.63196609
#>   0.226575533 -0.149817343  2.88677863
#>  -1.045795569  1.384826912  1.95525039
#>  -0.345514295 -2.456199022 -1.61123679
#>  -1.124163907 -0.579572244  1.80894242
#>  -1.399356893 -1.010240677 -3.60361510
#>  -1.464904649 -2.334497027 -2.41398232
#>  -2.761587838 -2.248864038  0.90344198
#>  -0.937764885 -0.083312210  1.88489312
#>   0.746064468  1.718620055  0.46436358
#>   1.281256847  1.312199593 -0.59907226
#>   1.259817374  2.891635402 -1.39522635
#>  -0.009724833 -2.928130277  1.12529741
#>   0.929118864 -2.585786295 -3.69108017
#>   1.195734363 -3.021432433 -4.10349518
#>  -1.002002297 -2.927119796  1.68649297
#>   1.619461618  2.136067838 -4.15674018
#>  -0.095495907 -2.477682083 -2.73907948
#>  -0.684953391  1.042105895 -2.25944397
#>   2.251559049  3.084420245 -0.17334693
#>  -0.421215016 -0.496248931  3.92061876
#>   1.089307504  1.682682104  1.00006912
#>   0.311286689 -0.037314335  1.59566879
#>   2.347670112 -0.074739463 -1.61985946
#>   0.426213171  0.081518522  2.32111816
#>   2.568246823  1.343652786  0.61085084
#>   0.751614411 -1.703696902 -3.41837550
#>  -2.026119478 -1.194183264  1.49301798
#>  -0.429442144  0.432267346  0.53404506
#>  -1.485614736 -2.418072566 -0.93133246
#>  -1.418415826 -1.918890934 -4.17573785
#>  -2.786756976 -0.833029874  2.89860661
#>   0.953528160  3.279050147  4.65139506
#>   3.036414152  1.447333882 -4.25572633
#>   4.171407368  3.926744710 -1.64357013
#>   2.609867870 -0.166898469  1.73970589
#>   3.258140685  1.590287747 -3.41818647
#>   0.164263189 -2.800671682 -1.75576067
#>   1.140920162 -2.207451923 -1.35999509
#>  -0.742445238 -1.041154033 -0.89993527
#>  -1.103641549 -5.109307428 -5.50785236
#>  -2.284013946 -3.456907597 -0.38296131
#>  -0.454434888  0.269226537 -1.82312826
#>  -1.045542573  1.769724843  4.47693880
#>   3.731610645  4.664361427 -2.91126237
#>  -1.038051168 -2.958470126  0.92022543
#>  -2.225043598 -1.996198129  1.06805516
#>   0.953332320  1.992210205  0.16002024
#>   0.525748299 -1.259804494 -2.47907049
#>  -1.197486501 -2.836832415 -0.05403530
#>  -1.533621787 -0.765384047  4.37233728
#>   3.096707945  4.379551525 -3.51970001
#>   2.709252265  1.606453410 -3.63832525
#>   2.255826075 -0.465203601 -3.10707907
#>  -0.190676567 -4.387830227 -0.75427658
#>   0.872517405 -1.633445189  0.10370934
#>   1.508505251  1.838468371  1.55981245
#>   1.598637329  0.361584329 -2.36487017
#>   0.615694603  0.254887404  0.54839784
#>   2.143894635  1.744226751 -2.83426202
#>   2.103102535  0.574926609 -2.59974863
#>  -0.554071627 -3.276310675  1.41268400
#>  -0.197332098  0.880778266 -5.01746104
#>  -2.406062427 -2.986327281  0.76170193
#>  -0.919887673  2.629815447  1.16584390
#>   1.815136313 -0.870707098 -0.91683061
#>   0.192720838 -1.344484362 -1.80393894
#>   1.435579341  0.981251372  0.85508358
#>   0.514882310  0.267316195  0.09238509
#>   0.408352770  0.729883045  3.70362352
#>  -0.159993532  2.154994424  0.49890372
#>  -0.409666097 -0.013286026 -0.65247642
#>  -1.889194126 -3.094411130  2.51156162
#>  -3.649793111 -0.248714079  2.22648318
#>   0.719825433  3.561403102  5.53225377
#>   3.046427654  4.610715465  0.26967560
#>   3.307160228  0.541563881 -1.95421789
#>   1.928707209 -1.541129136 -4.36617042
#>   0.067425505 -3.266373607  0.87280879
#>   0.199467228  0.614342683 -0.86902087
#>  -0.040980892  0.094096566 -2.13351223
#>   0.079090055 -0.793265853 -2.12554274
#>  -1.145685421 -2.145495306 -1.21936088
#>  -1.774545348 -2.279045379  3.33423684
#>   0.599144316 -0.574044521 -3.43954667
#>  -1.175064099 -5.082327176 -1.19799611
#>  -1.052790609  1.911143227  2.15869486
#>   2.027572160  3.611183415  0.12108359
#>   1.882404258  0.223640236  3.97114590
#>   2.949849534  2.535387654 -1.33124865
#>   1.886293126 -0.946268320 -1.02140264
#>   1.484923968  2.715203831 -2.71308074
#>   0.082438324 -0.858127416  3.30054546
#>  -1.177637831 -0.609853812 -1.16847884
#>  -2.152143192 -0.865889752  2.13270480
#>  -0.389184827  0.559302970 -0.54679802
#>   1.689813621  2.329188663 -0.08386881
#>   2.298431093  0.869289832  1.85395336
#>  -2.108541766 -1.989462099  3.21621739
#>  -0.652064706  1.260085974 -0.30266530
#>  -2.099972505  0.807434250  4.13949992
#>   0.188379418  0.154110264 -1.15638555
#>   0.055624106  1.673472461  1.40950611
#>   0.813281385  0.191928084 -0.05265852
#>   0.946788227  1.437513377  3.19530760
#>   0.629143309  1.614895998 -0.01242570
#>  -1.308356559 -1.021397961 -0.44327859
#>  -1.522933811 -0.352892953 -4.08141011
#>  -1.213608499  1.027846695  4.64313041
#>   2.159913641  3.276714311  0.08654678
#>   0.797404624 -1.623094786  0.55616417
#>  -1.604542558 -1.644984518  0.64793314
#>  -1.415518714 -0.173037712 -0.59181178
#>  -0.336009605 -0.021301984  3.17836156
#>   0.943665149  3.147794089 -1.04645049
#>   0.753013845 -0.298932783  1.10952946
#>   2.232249548  1.033007495 -0.27912571
#>   0.430187857 -0.848654277  3.55060955
#>   1.413999127  1.877508224 -1.02337357
#>   1.870337293  2.711116289 -3.85054255
#>   2.874427778 -0.271961941 -4.61564890
#>  -0.183429722 -1.185779845  5.29175690
#>   2.595906615  2.161883202 -4.20956530
#>   3.028075608  2.982272510 -4.08188637
#>   0.256853970 -2.651867232  0.12597845
#>   0.282682366 -1.291820101  2.00759873
#>   1.078119891  1.543928506 -1.10722498
#>  -1.022674202 -0.401080923  0.58404800
#>  -0.867413988  0.402097885 -3.36376170
#>  -0.493003961 -2.333957660  1.02261839
#>   1.345648239 -0.482462125  1.49174915
#>  -0.890795193  1.279684148 -3.42088442
#>  -2.258165669 -3.589929578  0.48880298
#>  -0.885710040  0.081871355  1.27651285
#>  -0.384949040 -1.049249632  1.47890640
#>  -0.603801512 -0.412589734  1.43889029
#>   0.643711025  1.730870109  0.42372458